% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_accessToken.R
\name{account.accessToken.create}
\alias{account.accessToken.create}
\title{Create an access token}
\usage{
account.accessToken.create(description, accessList, token, validFor = NULL)
}
\arguments{
\item{description}{Mandatory (string) a description for the token to be created}

\item{token}{Mandatory (string)}

\item{validFor}{Optional (int) number of seconds that the token should remain valid. If NULL token will remain valid forever.}

\item{list}{Mandatory (list, vector) a list of named lists. Each named list must contain the properties pathId and accessLevel}
}
\value{
...
}
\description{
Create an access token
}
