% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_accessToken.R
\name{account.accessToken.revoke}
\alias{account.accessToken.revoke}
\title{Revoke a personal access token}
\usage{
account.accessToken.revoke(accessTokenId, token)
}
\arguments{
\item{accessTokenId}{Mandatory (uuid), the id of the token to invalidate}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Revoke a personal access token
}
