% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{account.listRoot}
\alias{account.listRoot}
\title{Retrieve all layers and folders inside a specific folder}
\usage{
account.listRoot(
  rootName,
  token,
  pathTypes = NULL,
  pageStart = NULL,
  listAll = TRUE
)
}
\arguments{
\item{rootName}{a string representing the ID of the folder to be searched}

\item{token}{a string representing the login token}

\item{pathTypes}{a list representing the types of content that will be retrieved}

\item{pageStart}{a uuid representing the page on which the work starts}

\item{listAll}{a boolean controlling whether to retrieve only the first page or all pages}
}
\value{
a dictionary/json table containing the content that is retrieved
}
\description{
This function retrieves all layers and folders inside a specific folder,
and retrieves information about the folder itself.
}
