% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{\code{adjust} adjusts the structures returned by rejustify API:
structure, default values and keys.}
\usage{
adjust(block, id = NULL, column = NULL, items = NULL)
}
\description{
\code{adjust} adjusts the structures returned by rejustify API:
structure, default values and keys.
}
\examples{
to be added

}
