% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{attendance}
\alias{attendance}
\title{Attendance Dataset}
\format{
Data frame with columns
\describe{
\item{team}{Team City}
\item{team_name}{Team name}
\item{year}{Season year}
\item{total}{Total attendance across 17 weeks (1 week = no game)}
\item{home}{Home attendance}
\item{away}{Away attendance}
\item{week}{Week number (1-17)}
\item{weekly_attendance}{Weekly attendance number}
\item{full_name}{Combination of \code{team} variable and \code{team_name} variable}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2020/2020-02-04}
}
\usage{
attendance
}
\description{
A dataset that contains total attendance of NFL games per week and per team. It spans across multiple years from 2000 until 2019.
}
\keyword{datasets}
