% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_var_over_repetition.R
\name{average_ROF_facts}
\alias{average_ROF_facts}
\title{Plot average rate of forgetting over repetitions for each fact}
\usage{
average_ROF_facts(
  data,
  factNames = "factId",
  xlim = NULL,
  ylim = NULL,
  filepath = "../Figures"
)
}
\arguments{
\item{data}{A data frame. NA values will be removed before plotting.}

\item{factNames}{Column that shows the names of the facts (like factText or factAnswer). Default shows factId.}

\item{xlim}{A vector of 2 (for example: c(0, 10)), indicating the range of
the x-axis.If NULL the default value is used.}

\item{ylim}{A vector of 2 (for example: c(0, 1000)), indicating the range of
the y-axis.If NULL the default value is used.}

\item{filepath}{A relative or explicit path where plots will be saved}
}
\value{
Plot of Alpha over repetition, with every line a fact and a pdf file in filepath
}
\description{
Plots the average rate of forgetting (alpha) for each fact over the repetitions of
the facts.
}
\details{
Assumes that the data set has a repetition column with fact repetition in
integers and a column with the alpha's for all observations.

The legend is given in the order of the average alpha for the last repetition
of each fact. If there is more than one column (if there are too many facts
to fit in one column), the columns should be read one-by-one. The upper-left
value is the highest last alpha, the lower- right value is the lowest last
alpha. All last alpha's are also marker by a bigger marker in the plot. Be
aware that lines/markers may sometimes overlap.
}
\seealso{
Other average functions: 
\code{\link{average_ROF_participants}()},
\code{\link{average_RT_participants}()},
\code{\link{average_accuracy_participants}()}
}
\concept{average functions}
