% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_var_over_repetition.R
\name{average_RT_over_repetition}
\alias{average_RT_over_repetition}
\title{Plot average RT over repetitions}
\usage{
average_RT_over_repetition(
  data,
  xlim = NULL,
  ylim = NULL,
  filepath = "../Figures"
)
}
\arguments{
\item{data}{A data frame. NA values will be removed before plotting.}

\item{xlim}{A vector of 2 (for example: c(0, 10)), indicating the range of
the x-axis.If NULL the default value is used.}

\item{ylim}{A vector of 2 (for example: c(0, 1000)), indicating the range of
the y-axis.If NULL the default value is used: c(min(RT), mean(RT) +
(SD(RT))).}

\item{filepath}{A relative or explicit path where plots will be saved}
}
\value{
data frame
}
\description{
Plots the average RT of every session over the repetitions of the facts.
}
\details{
Assumes that the dataset has a repetition column with fact repetition in
integers
}
