% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagged_model}
\alias{bagged_model}
\title{Bagged Model Function}
\usage{
bagged_model(X, y, r.bagging, modelType)
}
\arguments{
\item{X}{Independent variables (predictors).}

\item{y}{Dependent variable (response).}

\item{r.bagging}{Number of bagging iterations.}

\item{modelType}{Type of machine learning model (linear, ridge, lasso, elastic net, random forest).}
}
\value{
If the response variable is factor, returns a list with fitted values and naive variables;
otherwise, returns fitted values.
}
\description{
This function performs bagging with different machine learning models.
}
