% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaggingInput.R
\name{bagging}
\alias{bagging}
\title{Bagging Model}
\usage{
bagging(X, y, model_type = "lasso", R = 10)
}
\arguments{
\item{X}{The dataset to train the bagging model on without the dependent variable}

\item{y}{The dependent variable of our dataset.}

\item{model_type}{The type of base model to be used for bagging. Options include "random_forest", "linear", "logistic", "lasso", "elastic_net", and "ridge".}

\item{R}{The number of bootstrap samples to be drawn.}
}
\value{
A list containing the average predictions, average coefficients (for penalized regression models), and naive importance scores (for penalized regression models).
}
\description{
The `bagging` function implements bagging, a model aggregation technique, to improve the predictive performance of a given model type.
}
\examples{
library(MASS)
data("Boston")
y <- Boston$medv
X <- Boston[, !names(Boston) \%in\% "medv"]
bagging(X, y, model_type = "lasso", R = 100)

}
