% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha_and_activation.R
\name{calculate_alpha_and_activation}
\alias{calculate_alpha_and_activation}
\title{Calculate fact activation and alpha (rate of forgetting)}
\usage{
calculate_alpha_and_activation(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Original data frame with columns for activation and alpha
}
\description{
Returns the data frame with activation and alpha for all trials. This will
override existing columns called 'alpha' or 'activation'.
}
