% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_means_by_category.R
\name{calculate_means}
\alias{calculate_means}
\title{Calculate means for each level of a categorical variable in a numeric vector}
\usage{
calculate_means(numeric_vector, categorical_vector)
}
\arguments{
\item{numeric_vector}{A numeric vector containing the data for which means are calculated.}

\item{categorical_vector}{A character vector containing the categorical variable.}
}
\value{
A named numeric vector with the means for each level of the categorical variable.
}
\description{
This function takes a numeric vector and a categorical vector as input. It calculates
the mean for each level of the categorical variable and returns a named numeric vector
with the means.
}
\examples{
# Example Data
numeric_data <- c(10, 15, 20, 25, 30, 35)
category_data <- c("A", "B", "A", "B", "A", "B")


# Calculate means for each category
calculate_means(numeric_data, category_data)

}
