% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_repetition.R
\name{calculate_repetition}
\alias{calculate_repetition}
\title{Calculate fact repetition}
\usage{
calculate_repetition(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Original data frame with column for repetition
}
\description{
Returns the data frame with a column for fact repetition for all trials. This will
override existing columns called 'repetition'.
}
\seealso{
Other calculation functions: 
\code{\link{calculate_alpha_and_activation}()}
}
\concept{calculation functions}
