% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{calculateRFM}
\alias{calculateRFM}
\title{calculateRFM}
\usage{
calculateRFM(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{weight_monetary}{}
}
\value{
RFM Score table
}
\description{
This function creates a RFM Model for a given Data
}
\details{
\code{data} Measure the recency, frequency, and monetary value of customer purchases.
Set a score to rank customers according to their purchase recency, frequency, and monetary value.
Calculate the overall RFM score. Analyze RFM group differences.
}
\examples{
calculateRFM(Data, 20, 20, 60)

}
