% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_space.R
\name{check_spaces}
\alias{check_spaces}
\title{Check gen3sis_spaces}
\usage{
check_spaces(spaces = NULL)
}
\arguments{
\item{spaces}{either a gen3sis_spaces object to be checked, or NULL}
}
\value{
If a spaces object is provided, either a stop with printed error
report or a pass statement.
If spaces=NULL, this function returns lists of accepted values categories
according to \code{check_spaces()}
}
\description{
Check gen3sis_spaces
}
\examples{
print("TODO")
}
