% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circaPower.R
\name{circaPower}
\alias{circaPower}
\title{Analytical circadian power calculation based on F-statistics}
\usage{
circaPower(n, A, sigma, phi = 0, period = 24, cts = NULL, alpha = 0.05)
}
\arguments{
\item{n}{Sample size.}

\item{A}{Amplitude of the sinusoidal curve: \eqn{A * sin(2\pi/period * (phi + cts)) + M}.}

\item{sigma}{Standard deviation \eqn{\sigma} in the independent Normal error term \eqn{N(0,\sigma)}.}

\item{phi}{Phase shift of the sinusoidal curve. Default is 0.}

\item{period}{Period of the sinusoidal curve. Default is 24.}

\item{cts}{Circadian times of samples.}

\item{alpha}{Type I error control. Default is 0.05.}
}
\value{
Statistical power calculated from a closed-form formula based on F-statistics.
}
\description{
Analytical circadian power calculation based on F-statistics in the cosiner model.
}
\details{
Analytically calculate power of omics circaidan data under indepednent Normal error assumption using F-statistics.
}
\examples{
n = 100
cts = seq(0,24,length.out = n+1)[-1]
A = 1
sigma = 1
phi = 0
circaPower(n, A, sigma, phi=0, period = 24, cts=cts, alpha = 0.05)
}
\author{
Wei Zong, Zhiguang Huo
}
