% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_avg.R
\name{coal_prod}
\alias{coal_prod}
\title{Calculate the average Coal production.}
\usage{
coal_prod(x)
}
\arguments{
\item{x}{A numerical value between 1980 and 2021.}
}
\value{
A numerical value which is the average value of the total coal production
for that year if x is between 1980 and 2021 else it will return an error message
}
\description{
This function takes an input between 1980 and 2021 and returns the
average value of the coal production for the given year.
}
\examples{
coal_prod(2005)
coal_prod(1990)
coal_prod(1980)


}
