% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color_richness}
\alias{color_richness}
\title{define gen3sis richness color scale}
\usage{
color_richness(n)
}
\value{
return a colorRampPalette with gen3sis richness collors
}
\description{
define gen3sis richness color scale
}
