% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color_richness_non_CVDCBP}
\alias{color_richness_non_CVDCBP}
\title{Define gen3sis richness color scale for non colour-vision deficient and colour-blind people}
\usage{
color_richness_non_CVDCBP(n)
}
\arguments{
\item{n}{corresponds to the \link{colorRampPalette} parameter}
}
\value{
returns a \link{colorRampPalette} function with the gen3sis richness colors
}
\description{
Define gen3sis richness color scale for non colour-vision deficient and colour-blind people
}
