% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{Cons_variance

This function is used to do the Equal variance diagnosis of the linear model.}
\alias{Cons_variance

This function is used to do the Equal variance diagnosis of the linear model.}
\alias{Cons_variance}
\title{Equal variance diagnosis}
\usage{
Cons_variance(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector of outcome values (should be a continuous variable).}
}
\value{
The plot of the residuals versus fitted Y.
}
\description{
Equal variance diagnosis
}
\examples{
data(mtcars)
Cons_variance(X= cbind(mtcars$cyl, mtcars$hp, mtcars$drat, mtcars$wt), Y = mtcars$mpg)

}
