% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{create_config}
\alias{create_config}
\title{create_config creates eiter an empty configuration oder prepopulates a configuration from a config file}
\usage{
create_config(config_file = NA)
}
\arguments{
\item{config_file}{NA the create an empty config or the path to a valid configuration file}
}
\description{
create_config creates eiter an empty configuration oder prepopulates a configuration from a config file
}
