% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_species.R
\name{create_initial_species}
\alias{create_initial_species}
\title{create_initial_species allows the user to populate the world at the beginning of a simulation}
\usage{
create_initial_species(landscape, config)
}
\arguments{
\item{landscape}{the current landscape}

\item{config}{the current config}
}
\value{
a list of species
}
\description{
create_initial_species allows the user to populate the world at the beginning of a simulation
}
\details{
The user can create any number of species in this function. For every species she shall choose a number of
habitable cells from the landscape and call 'create_species'. Afterwards she must initialize the species[["traits"]]
matrix with the desired initial traits values.
}
