% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_creation.R
\name{create_input}
\alias{create_input}
\title{Create an landscape input}
\usage{
create_input(
  landscapes,
  cost_function,
  directions,
  output_directory,
  timesteps = NULL,
  habitability_masks = NULL,
  calculate_full_distance_matrices = F,
  crs = NULL,
  overwrite_output = F,
  verbose = F
)
}
\arguments{
\item{landscapes}{list of named list(s) of raster(s) or raster file(s) name(s). Starting from the present towards the past.
The list names are important since these are the environmental names}

\item{cost_function}{function that returns a cost value between a pair of cell (neighbours) that should have the following signature:
#' cost_fuinction <- function(src, src_habitable, dest, dest_habitable){
 #rules for envoriolmental factors to be considered (e.g. elevation)
 #return(cost value)
}
where src is a vector of environmental conditions for the origin cell, src_habitable (TRUE or FALSE) for habitable condition of the origin cell, dest is a vector of environmental conditions for the destination cell, dest_habitable  (TRUE or FALSE) for habitable condition of the destination cell}

\item{directions}{4, 8, 16 oder adjacency matrix (see gdistance package)}

\item{output_directory}{path for storing the gen3sis ready landscape}

\item{timesteps}{vector of names for every tiumestep to represent the timestep at gen3sis ready landscape. if NULL, timsteps are sequentialy numbered from 0 (present)}

\item{habitability_masks}{either (1) a list of files or rasters starting from the present indicating the habitability of cell at that timestep.
#' or (2) NULL, every cell with at least one NA as environmentalvalue(s) will be considered inhabitable}

\item{calculate_full_distance_matrices}{TRUE or FALSE. If TRUE calculates the entire distance matrix for every timestep and between all habitable cells (faster CPU time, higher storage required). If FALSE, only local distances are calculated (slower CPU time when simulating but smaller gen3sis landscape size)}

\item{crs}{the coordinate reference system in crs format (see rater::crs)}

\item{overwrite_output}{TRUE or FALSE}

\item{verbose}{print distance calculation progress}
}
\description{
Create an landscape input
}
\examples{
##TODO##

# add example
# create_input(
#   landscapes,
#   cost_function,
#   directions,
#   output_directory,
#   timesteps = NULL,
#   habitability_masks = NULL,
#   calculate_full_distance_matrices = F,
#   crs = NULL,
#   overwrite_output = F,
#   verbose = F
# )
}
