% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_species.R
\name{create_species}
\alias{create_species}
\title{The function create_species creates a new species.}
\usage{
create_species(id, initial_cells, config)
}
\arguments{
\item{id}{the id for the new species, might be overriden, tbd}

\item{initial_cells}{a list of initial cells (strings) to occupy}

\item{config}{the configuration information}
}
\value{
returns a newly created species occupying the provided initial cells
}
\description{
The function create_species creates a new species.
}
\details{
This function is intented to be used in the create_ancestor_species function at the configuration 
of a simulation. It will create a species object representing one species in the simulation occupying the 
given list of initial cells.
}
\examples{
##TODO##

# add example
#create_species This creates a new species

# create_species(id, initial_cells, config)
}
