% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{CreateModelList}
\alias{CreateModelList}
\title{Create Model List}
\usage{
CreateModelList(...)
}
\arguments{
\item{...}{All objects containing regression models to be put into a list.}
}
\value{
A list containing all regression models from user input.
}
\description{
A generator function that creates a list of regression models assigned to objects that were predefined by user.
}
\examples{
m1 <- lm(lifeExp ~ year + pop + gdpPercap, gapminder::gapminder)
m2 <- lm(lifeExp ~ year, gapminder::gapminder)
mList <- CreateModelList(m1, m2)


}
