% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_suffix.R
\name{csv_suffix}
\alias{csv_suffix}
\title{Append .csv to string}
\usage{
csv_suffix(x)
}
\arguments{
\item{x}{string. Usually intended to be the file name}
}
\value{
The original string (x) with ".csv" appended.
}
\description{
Append .csv to string
}
\examples{
csv_suffix(filename)
}
