% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_Chen.R
\docType{data}
\name{cts_Chen}
\alias{cts_Chen}
\title{Circadian Times of Brodmann’s Areas (BA) 11 and 47 Brain Aging Data.}
\format{
A circadian time vector.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71620}
}
\usage{
cts_Chen
}
\description{
Observed sample time-of-death from human postmortem brain data (Brodmann's area 11 and 47 in the prefrontal cortex).
}
\references{
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
