% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_kernel_generator.R
\name{cts_kernel_generator}
\alias{cts_kernel_generator}
\title{Circadian Times Sampling from the Kernal Density estimated from Real Data}
\usage{
cts_kernel_generator(observed_cts, cts_lb = 0, cts_ub = 24, n)
}
\arguments{
\item{observed_cts}{Observed circadian times in real data. Lie between cts_lb and cts_ub.}

\item{cts_lb}{Lower bound of circadian times.}

\item{cts_ub}{Lower bound ofcircadian times.}

\item{n}{Number of time of death to generate.}
}
\value{
A circadian times vector of length n
}
\description{
Circadian times sampled from the density estimated from observed sampling times in real data.
}
\details{
Generate circadian times at different sample sizes from kernel density estimated
from observed sampling times in real data.
}
\examples{
data(cts_Chen)
cts_kernel_generator(observed_cts=cts_Chen, cts_lb = 0, cts_ub = 24, n = 100)
}
\author{
Wei Zong, Zhiguang Huo
}
