% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_Ketchesin.R
\docType{data}
\name{cts_Ketchesin}
\alias{cts_Ketchesin}
\title{Circadian Times of Dorsal and Ventral Straitum Data (control subjects).}
\format{
A circadian time vector.
}
\usage{
cts_Ketchesin
}
\description{
Observed sample time-of-death from human postmortem brain data (dorsal and ventral straitum data).
}
\references{
Ketchesin, K. D., Zong, W., Hildebrand, M. A., Seney, M. L., Cahill, K. M., Scott, M. R., ... & McClung, C. A. (2021). Diurnal rhythms across the human dorsal and ventral striatum. Proceedings of the National Academy of Sciences, 118(2).
}
\keyword{datasets}
