% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_Seney.R
\docType{data}
\name{cts_Seney}
\alias{cts_Seney}
\title{Circadian Times of Dorsolateral Prefrontal Cortex Data (control subjects)}
\format{
A circadian time vector.
}
\usage{
cts_Seney
}
\description{
Observed sample time-of-death from human postmortem brain data (dorsolateral prefrontal cortex).
}
\references{
Seney, M. L., Cahill, K., Enwright, J. F., Logan, R. W., Huo, Z., Zong, W., ... & McClung, C. A. (2019). Diurnal rhythms in gene expression in the prefrontal cortex in schizophrenia. Nature communications, 10(1), 1-11.
}
\keyword{datasets}
