% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_stats.R
\name{dataset_stats}
\alias{dataset_stats}
\title{Descriptive statistics about the data set}
\usage{
dataset_stats(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Three data frames ($general, $lessons, $participants)
}
\description{
\code{dataset_stats} describes several statistics regarding the data set,
these are: Number of participants, number of unique facts, number of lessons,
participants per lesson, lessons per participant and sessions per
participant.
}
\details{
The number of participants, number of unique facts, number of lessons and
participants per lesson are always printed in the console. The data frame
with the sessions and lessons per participants is not printed, since the
amount of participants can be substantial. All information can be found in
$general, $lessons and $participants.
}
