% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diametre.R
\name{diameter}
\alias{diameter}
\title{Pipe diameter}
\usage{
diameter(Q, V)
}
\arguments{
\item{Q}{volumetric flow rate}

\item{V}{fluid velocity}
}
\value{
the diameter of the conduit through wich the fluid will flow
}
\description{
Determine the ideal pipe diameter given the desired flow rate and velocity
}
\examples{
diameter(2,20)
7.136496

}
