% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculate the shortest path from the initial node to every node in the graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with three variables`v1`,`v2`and `w`.
*`v1`,`V2`:A pair of nodes in the graph.
*`w`:The weight of the edge from v1 to v2.}

\item{init_node}{A numeric scalar that exist in the graph.}
}
\value{
new_dist A vector that shows the shortest path to every other node 
  from the staring point.
}
\description{
The function uses Dijkstras algorithm to Calculate the shortest path from
  the initial node to every node in the graph.
}
\examples{
  wiki_graph <- data.frame(
  v1 = c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2 = c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w = c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
  )
  shortest_paths <- dijkstra(wiki_graph, 3)

}
\seealso{
[https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm]for more 
  information about Dijkstras algorithm.
}
