% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityDynamics.r
\name{divDyn}
\alias{divDyn}
\title{Time series from metrics of diversity dynamics}
\usage{
divDyn(dat, tax = "genus", bin = "bin", breaks = NULL,
  coll = "collection_no", ref = "reference_no", om = NULL,
  noNAStart = F, inf = F, data.frame = T, filterNA = FALSE)
}
\arguments{
\item{dat}{(data.frame): the data frame of fossil occurrences.}

\item{tax}{(character): variable  name of the occurring taxa (variable type: factor) - such as "occurrence.genus_name"}

\item{bin}{(character): variable name of the time slice numbers of the particular occurrences. This variable should be numeric and should increase as time passes by (use negative values for age estimates!).}

\item{breaks}{(numeric): If NULL (default) the used values in the 'bin' variable will designate independent time slices that follow each other in succession. In case of positive integer bin identifiers, and if noNAStart=FALSE,  the index of the row will be the bin number. If a vector is provided, than the numeric entries in 'bin' will be binned similarly to the hist() function. The order of elements in this vector is arbitrary.}

\item{coll}{(character): the variable name of the collection identifiers. (optional, only for filtering!)}

\item{ref}{(character): the variable name of the reference identifiers. (optional, only for filtering!)}

\item{om}{(character): the om parameter of the omit() function. If set to NULL (default), then no occurrences will be omitted before the execution of the function.}

\item{noNAStart}{(logical): useful when the dataset does not start from bin No. 1, but positive integer bin numbers are provided. Then noNAStart=TRUE will cut the first part of the resulting table, so the first row will contain the estimates for the lowest bin number.}

\item{inf}{(logical): should infinites be converted to NAs?}

\item{data.frame}{(logical): should be a data frame or a matrix?}

\item{filterNA}{(logical): the filterNA() parameter of the omit function.}
}
\description{
This function calculates 
various metrics in the form of time series from resolved PBDB occurrences.
}
\details{
The following variables are produced:

bin: the time slice number, or the numeric identifier of the time slice

tThrough: the number of through-ranging taxa

tOri: the number of originating taxa

tExt: the number of taxa getting extinct

tSing: the number of stratigraphic singleton (single-interval) taxa

t2d: the number of taxa that are present in the i-1th and the ith interval (lower two timers)

t2u: the number of taxa that are present in the ith and the i+1th interval (upper two timers)

tGFu: upper gap-fillers ($later)

tGFd: lower gap-fillers ($later)

t3: the number of three timer taxa, present in time slice i-1, i, and i+1

tPart: the part timer taxa, present in time slice i-1,and i+1, but not in i

extProp: proportional extinctions, including single-interval taxa

oriProp: proportional originations, including single-interval taxa

extPC: the per capita extinction rates of foote (not normalized with interval length!)

oriPC: the per capita origination rates of foote (not normalized with interval length!)

ext3t: the three-timer extinction rates

ori3t: the three-timer origination rates

extC3t: the corrected three-timer extinction rates

oriC3t: the corrected three-timer origination rates

divSIB: sampled-in-bin diversity (richness)

divCSIB: corrected sampled-in-bin diversity (richness)

divBC: boundary-crosser diversity (richness)

divRT: range-through diversity (richness)

sampRange: sampling probability (Foote)

samp3t: three-timer sampling completeness (used as a correcting value in extC3t, oriC3t and divCSIB)

extGF: gap-filler extinction rates (Alroy, 2014)

oriGF: gap-filler origination rates (Alroy, 2014)

E2f3: second-for-third extinction propotions (Alroy, 2015)

O2f3: second-for-third origination propotions (Alroy, 2015)

ext2f3: second-for-third extinction rates (based on Alroy, 2015)

ori2f3: second-for-third origination rates (based on Alroy, 2015)

References:

Foote, M. (2000) Origination and Extinction Components of Taxonomic Diversity: General Problems. Paleobiology 26, 74-102. doi:10.1666/0094-8373(2000)26[74:OAECOT]2.0.CO;2).

Alroy, J. (2008) Dynamics of origination and extinction in the marine fossil record. Proceedings of the National Academy of Science 105, 11536-11542. doi: 10.1073/pnas.0802597105

Alroy, J. (2014) Accurate and precise estimates of origination and extinction rates. Paleobiology 40, 374-397. doi: 10.1666/13036

Alroy, J. (2015) A more precise speciation and extinction rate estimator. Paleobiology 41, 633–639. doi: 10.1017/pab.2015.26
}
\examples{
# import data
  data(corals)
  data(stages)

# calculate metrics of diversity dynamics
   dd <- divDyn(corals, tax="genus", bin="slc")

# plotting
  plotTS(stages, shading="series", boxes="per", xlim=c(260,0), 
    ylab="range-through diversity (genera)", ylim=c(0,230))
  lines(stages$mid, dd$divRT, lwd=2)

 # with omission of single reference taxa  
   ddNoSing <- divDyn(corals, tax="genus", bin="slc", om="ref")
   lines(stages$mid, ddNoSing$divRT, lwd=2, col="red")

 # using the estimated ages (less robust) - 10 million years
   # mean ages (should be negative to maintain order)
   corals$me_ma <- -apply(corals[, c("max_ma", "min_ma")], 1, mean)
   # divDyn
   ddRadio10 <- divDyn(corals, tax="genus", bin="me_ma", breaks=seq(0,-250,-10))
   lines(-ddRadio10$bin, ddRadio10$divRT, lwd=2, col="green")
      
 # legend
   legend("topleft", legend=c("all", "no single-ref. taxa", "all, estimated ages"), 
     col=c("black", "red", "green"), lwd=c(2,2,2), bg="white")
   

}
