% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupsBetweenGroups.R
\name{dupsBetweenGroups}
\alias{dupsBetweenGroups}
\title{A dupsBetweenGroups Function}
\usage{
dupsBetweenGroups(FULL, Coder)
}
\description{
This function allows you to find the duplicated rows basing on the common column they have.
}
\examples{
dupsBetweenGroups()
}

