% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic net.R
\name{elastic_net_regression}
\alias{elastic_net_regression}
\title{Elastic Net Regression with Cross-Validated Alpha Selection}
\usage{
elastic_net_regression(
  X,
  y,
  alphas = NULL,
  lambdas = NULL,
  method = "glmnet",
  K = NULL,
  ...
)
}
\arguments{
\item{X}{Matrix of predictors (independent variables), can be a mix of continuous, discrete, and binary predictors}

\item{y}{Response variable (binary or continuous)}

\item{alphas}{Vector of alpha values to try}

\item{lambdas}{Regularization parameter}

\item{method}{Method for regression ('glmnet')}

\item{K}{Number of top predictors to include in the model}

\item{...}{Additional arguments to be passed to the regression method}
}
\value{
Fitted model object
}
\description{
Fits an elastic net regression model with cross-validated alpha selection.
}
\examples{
X <- {}
y <- {}
model <- elastic_net_regression(X, y, method = 'glmnet', K = 10)
}
