% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_predict}
\alias{ensemble_predict}
\title{Ensemble Learning Prediction}
\usage{
ensemble_predict(X, y)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}
}
\value{
A list containing the predictions and accuracies of the individual models and the ensemble model.
- ensembled_prediction: The ensemble prediction.
- elastic.net: List containing information about the Elastic Net model, including predicted values.
- random.forest: List containing information about the Random Forest model, including predicted values.
Example usage:
X <- matrix(rnorm(1000 * 2), ncol = 2)
y <- rnorm(1000)
result <- ensemble_predict(X, y)
}
\description{
This function performs ensemble learning using Random Forest and Elastic Net
as base learners. It trains on 70\% of the data and tests on the remaining 30\%. The function returns predictions
and accuracies for individual models as well as the ensemble model using a simple mean approach.
}
