% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleinput.R
\name{ensemble_prediction}
\alias{ensemble_prediction}
\title{ensemble prediction}
\usage{
ensemble_prediction(X, y, models, weights = NULL)
}
\arguments{
\item{X}{The dataset to train the ensemble prediction model on without the dependent variable}

\item{y}{The dependent variable of our dataset.}

\item{models}{A character vector specifying the types of base models to be used for ensemble prediction.}

\item{weights}{Optional weights for combining predictions from different models. If not provided, equal weights are used.}

\item{output}{To exit from the console return "4"}
}
\value{
A list containing the weighted predictions, predictions from individual models, and model coefficients.
}
\description{
The `ensemble_prediction` function implements ensemble prediction, combining predictions from multiple models with optional weights.
}
\examples{
library(MASS)
data("Boston")
y <- Boston$medv
X <- Boston[, !names(Boston) \%in\% "medv"]
ensemble_prediction(X, y, models = c("linear", "lasso", "random_forest"))


}
