% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the greatest common divisor of two numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x, y}{A pair of input integers.}
}
\value{
r_2 The greatest common divisor of two numbers.
}
\description{
The function uses Euclidean algorithm to find the greatest common divisor 
  of two numbers.
}
\examples{
euclidean(123612,13892347912)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm} for more 
  information about Euclidean algorithm.
}
