% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example function that prints Hello world!}
\usage{
example_function(x)
}
\arguments{
\item{x}{Integer (1 = Hello World message)}
}
\value{
prints a 'Hello world!' string
}
\description{
Example function that prints Hello world!
}
\examples{
x <- 1
example_function(x)

y <- 0
example_function(y)
}
