% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplefunction.R
\name{examplefunction}
\alias{examplefunction}
\title{Example function}
\usage{
examplefunction(x)
}
\arguments{
\item{x}{Integer of value 1}
}
\value{
prints a 'Hello world!' string
}
\description{
Example function
}
\examples{
x <- 1
examplefunction(x)
}
