% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_column_conversion.R
\name{excel_column_conversion}
\alias{excel_column_conversion}
\title{Converts Excel column references}
\usage{
excel_column_conversion(col)
}
\arguments{
\item{col}{Column number to turn into Excel's 'A1 reference style' for columns. Alternatively, an 'A1 reference style' string to turn into a column number.}
}
\value{
A string with the corresponding 'A1 reference style' column reference. Alternatively, a number corresponding to the given 'A1 reference style' column reference.
}
\description{
Converts Excel 'A1 reference style' column references to 'numeric' or vice versa.
}
\examples{
excel_column_conversion(1)
excel_column_conversion(26)
excel_column_conversion(27)
excel_column_conversion("A")
excel_column_conversion("Z")
excel_column_conversion("AA")
}
