% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frottements.R
\name{f}
\alias{f}
\title{frictions}
\usage{
f(Re)
}
\arguments{
\item{Re}{Reynolds number}
}
\value{
The friction factor f
}
\description{
Frictions in pipe flow refers to the viscous interaction
between the mooving fluid and the pipe wall. It depend of the flow regime.
}
\examples{
f(2000)
0.032

f(3500)
f(3500)
Enter the roughness of the conduit: 0.2
enter the pipe diameter: 0.010
0.008383059

}
