% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLad.r
\name{fadLad}
\alias{fadLad}
\title{FAD - LAD matrix from occurrence data}
\usage{
fadLad(dat, tax, bin, maxdate = NULL, mindate = NULL)
}
\arguments{
\item{dat}{\code{(data.frame)} Occurrence data.}

\item{tax}{\code{(character)} The column name of taxon names.}

\item{bin}{\code{(character)} The column name of bin names, has to be an integer}

\item{maxdate}{\code{(character)} Optional, the column names of maximum estimated ages of occurrences.}

\item{mindate}{\code{(character)} Optional, the column names of minimum estimated ages of occurrences.}
}
\description{
Function to generate range data from an occurrence dataset.
}
\details{
The function will output First and Last Appearance Data of the taxa in the dataset. Keep in mind that incomplete sampling will influence these data and will make the ranges appear shrunken.

The following variables are produced:

attribute \code{row.names}: The names of the taxa.

\code{FAD}: First appearance data, given in time slice numbers.

\code{LAD}: Last appearance data, given in time slice numbers.

\code{range}: the range of the taxo in the given time slices

\code{myFAD}: optionally, first appearance dates in numeric ages.

\code{myLAD}: optionally, last appearance dates in numeric ages.

\code{duration}: optionally, the duration of taxa in numeric ages. It is given for single-interval taxa as well assuming that the taxa's range span over the entire time slice.
}
\examples{
data(corals)
fl <- fadLad(corals, tax="genus", bin="slc", maxdate="max_ma", mindate="min_ma")
}
