% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{\code{fill} submits the dataset and structure to the fill API endpoint
to retrieve the desired extra data points. At the current stage
dataset must be rectangular, and structure should be in the shape proposed
analyze function.}
\usage{
fill(
  df = NULL,
  structure = NULL,
  keys = NULL,
  default = NULL,
  shape = "vertical",
  inits = 1,
  sep = ",",
  db = TRUE,
  accu = 0.75,
  form = "full",
  token = getOption("rejustify.token"),
  email = getOption("rejustify.email"),
  url = getOption("rejustify.mainUrl")
)
}
\description{
\code{fill} submits the dataset and structure to the fill API endpoint
to retrieve the desired extra data points. At the current stage
dataset must be rectangular, and structure should be in the shape proposed
analyze function.
}
\examples{
fill(df, structure)

}
