% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{Fizz Buzz a vector.}
\usage{
fizzbuzz(n)
}
\arguments{
\item{n}{an int vector}
}
\value{
vector of original int values and Fizz's and Buzz's
}
\description{
This function takes in an integer vector and returns an output vector of the input vector,
but any integer divisible by 3 gets replaces with "Fizz" and any integer divisible by 5
gets replaced with "Buzz", where being divisible by 5 takes precedence
}
\examples{
fizzbuzz(c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30))
fizzbuzz(c(100,200,300,400,500))
}
