% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff weather data}
\format{
A data frame with 25,449 observations with 5 columns.
\describe{
\item{Date}{The dates of each individual observation; Unit: date;       Datatype: chr}
\item{PRCP}{The amount of percipitation on each date; Unit: inch;       Datatype: num}
\item{SNOW}{The amount of snow on each date;          Unit: inch;       Datatype: num}
\item{TMAX}{Temperature maximum of each date;         Unit: Fahrenheit; Datatype: int}
\item{TMIN}{Temperature minimum of each date;         Unit: Fahrenheit; Datatype: int}
}
}
\usage{
Flagstaff_Weather
}
\description{
A set of data about Flagstaff's weather from 1950 to 2019.
}
\keyword{datasets}
