% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debit.R
\name{flow}
\alias{flow}
\title{Fluid flow}
\usage{
flow(Re, rho, nu, D)
}
\arguments{
\item{Re}{Reynolds number}

\item{rho}{fluid density (kg/m3)}

\item{nu}{dynamic viscosity (Pa.s)}

\item{D}{caracteristic length (m)}
}
\value{
The fluid flow
}
\description{
Calculate the flow rate of fluid flowing in a pipe.
}
\examples{

flow(1000,2000,0.1,0.001)
3.926991e-05

}
