% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_avg.R
\name{gas_prod}
\alias{gas_prod}
\title{Calculate the average Natural Gas production.}
\usage{
gas_prod(x)
}
\arguments{
\item{x}{A numerical value between 1980 and 2021.}
}
\value{
A numerical value which is the average value of the total gas production
for that year if x is between 1980 and 2021 else it will return an error message
}
\description{
This function takes an input between 1980 and 2021 and returns the
average value of the gas production for the given year.
}
\examples{
gas_prod(2005)
gas_prod(1990)
gas_prod(1980)


}
