% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_main.R
\docType{package}
\name{gen3sis}
\alias{gen3sis}
\title{Gen3sis: General Engine for Eco-Evolutionary Simulations}
\description{
Gen3sis: General Engine for Eco-Evolutionary Simulations
}
\examples{
\dontrun{
#runing a Gen3sis simulation
# add(1, 1)
# add(10, 1)
}
}
