% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{get_dispersal_values}
\alias{get_dispersal_values}
\title{get_dispersal_values is a user provided function that creates a number of dispersal values for a given species.}
\usage{
get_dispersal_values(num_draws, species, landscape, config)
}
\arguments{
\item{num_draws}{'num_draws' the expected number of dispersal values}

\item{species}{'species' the current species for which the values are to be produced}

\item{landscape}{'landscape' the current landscape}

\item{config}{'config' the current config}
}
\value{
a numerical vector of length num_draws with dispersal values
}
\description{
get_dispersal_values is a user provided function that creates a number of dispersal values for a given species.
}
\details{
These dispersal values are used for two different tasks. They are used to evaluate pair-wise dispersal
events between colonozed and uninhabited cells. They are furthermore used during the geographical clustering of
species populations when determining which cells are in range of each other and belong together.
Note: if the distances are randomized the cluster-formation may be asymetrical. Therefore the ordering of all
clustering operations is randomized
}
