% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciation.R
\name{get_divergence_factor}
\alias{get_divergence_factor}
\title{get_divergence_factor gives a value or matrix indicating the divergence
between geographically separated clusters/populations of a species}
\usage{
get_divergence_factor(species, cluster_indices, landscape, config)
}
\arguments{
\item{species}{the current species}

\item{cluster_indices}{an index vector indicating the cluster every occupied cell is part of}

\item{landscape}{the current landscape}

\item{config}{the current config}
}
\value{
a single value or a matrix of divergences between all clusters occuring in clusters_indices
}
\description{
get_divergence_factor gives a value or matrix indicating the divergence
between geographically separated clusters/populations of a species
}
\details{
this function determines the increase in deivergence between separated clusters of a species.
It can either return a single value or a full cluster by cluster matrix. For every timestep two cells
belong to the same cluster their divergence is reduced by 1.
Therefore the divergence values should be scaled as rewuired.
}
