% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_place_point.R
\name{get_place_point}
\alias{get_place_point}
\title{get_place_point}
\usage{
get_place_point(
  place_name,
  place_name_detail = NULL,
  place_value,
  closest_match = FALSE,
  filter_extra_cols = TRUE,
  ignore_multiple_output_warnings = FALSE
)
}
\arguments{
\item{place_name}{string containing name of the desired place}

\item{place_name_detail}{optional additional identifying string. Usually used to specify country.}

\item{place_value}{required string designating the place value, e.g. city, country, town. See: https://wiki.openstreetmap.org/wiki/Key:place}

\item{closest_match}{if there are multiple results, computes Levenshtein distance and filters to just the closest match.}

\item{filter_extra_cols}{default TRUE. OpenStreetMap nodes contain a variety of additional data such as names in different languages. `filter_extra_cols` filters these to just osm_id, name, english name, and place value.}

\item{ignore_multiple_output_warnings}{ignores warnings for multiple results if the `place_name` and `place_extra_detail` arguments return more than one valid point.}
}
\value{
"sf" "data.frame"
}
\description{
Queries OpenStreetMap to find the geographic point for a specified place
}
\examples{
## Not run:

# Multiple results, Sydney, Australia & Sydney, Canada
get_place_point("Sydney", place_value = "city")

# Since these are in different countries we can narrow down using the `place_extra_detail` argument: 
get_place_point("Sydney", "Australia", place_value = "city")


# Multiple results, Taipei, Taiwan & New Taipei, Taiwan
get_place_point("Taipei", place_value = "city")

# Since these are in the same country we should use `closest_match = TRUE` to match with only 'Taipei'
get_place_point("Taipei", place_value = "city", closest_match = TRUE)
# If 'New Taipei' is desired then just directly search for it: 
get_place_point("New Taipei", place_value = "city")

## End(Not run)
}
