% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{getValuesAlongLine}
\alias{getValuesAlongLine}
\title{Request to obtain the raster value for each point along a line}
\usage{
getValuesAlongLine(
  pathId,
  timestampId,
  line,
  token = NULL,
  epsg = 4326,
  asRaster = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{line}{Mandatory (sf geometry) of type line in WGS84}

\item{token}{Optional (string)}

\item{epsg}{Optional (int) default 4326}

\item{asRaster}{Optional (logical) default FALSE, if TRUE, an R raster object is returned instead of a matrix}
}
\value{
A named list with a "raster" property of type Raster, or Matrix depending on asRaster parameter containing the raster for points along a line
}
\description{
Request to obtain the raster value for each point along a line
}
