% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{Independence}
\alias{Independence}
\title{Independence}
\usage{
Independence(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector of outcome values (should be a continuous variable).}
}
\value{
The autocorrelation plots of the residuals.
}
\description{
This function is used to do the independence diagnosis of the linear model.
}
\examples{
data(mtcars)
Independence(X= cbind(mtcars$cyl, mtcars$hp, mtcars$drat, mtcars$wt), Y = mtcars$mpg)

}
