% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexify.R
\name{indexify}
\alias{indexify}
\title{Convert vector into an index}
\usage{
indexify(vector, index_to, index_to_hundred = TRUE, rounding = 2)
}
\arguments{
\item{vector}{Vector or numeric column to convert to an index.}

\item{index_to}{A number, usually an element in vector, corresponding to the base period.}

\item{index_to_hundred}{Default TRUE. Multiplies index by 100, i.e. base period will be 100 instead of 1. Any non-TRUE value will not multiply the index by 100.}

\item{rounding}{Default 2. See base function round, digits argument.}
}
\value{
A vector derived by indexing vector to the index_to value.
}
\description{
Convert vector into an index
}
\examples{
indexify(c(1:100), 1)
indexify(c(1:100), 1, FALSE)
indexify(c(1:100), c(1:100)[40])
indexify(c(1:100), c(1:100)[40], FALSE)
indexify(c(1:100), c(1:100)[40], FALSE, 5)
}
