% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_RT.R
\name{individual_RT}
\alias{individual_RT}
\title{Plot RT for all Facts for all Participants over Time}
\usage{
individual_RT(
  data,
  sessionId = NULL,
  normalizeTime = FALSE,
  xlim = NULL,
  ylim = NULL,
  filepath = "../Figures"
)
}
\arguments{
\item{data}{A data frame. NA values will be removed before plotting.}

\item{sessionId}{Provide a single sessionId string to plot that session. If
sessionId is NULL all sessions will be plotted.}

\item{normalizeTime}{If TRUE, the times of all facts will be normalized (they
will start at 0). If FALSE, the times will not be normalized and data
points will occur relative to their occurrence during the session.}

\item{xlim}{A vector of 2 (for example: c(0, 10)), indicating the range of
the x-axis.If NULL the default value is used: c(0, z). Where z is the max
time.}

\item{ylim}{A vector of 2 (for example: c(0, 10)), indicating the range of
the y-axis.If NULL the default value is used: c(a, b). Where a is the
minimum reaction time and b is the average reaction time plus the standard
deviation of the reaction time.}

\item{filepath}{A relative or explicit path where plots will be saved}
}
\value{
data frame
}
\description{
\code{individual_RT} plots the RT of one session over time. Incorrect
answers are denotes with a red marker.
}
