% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMissing}
\alias{isMissing}
\title{checks if a variable has non-missing values}
\usage{
isMissing(x)
}
\arguments{
\item{x}{Variable to test.}
}
\description{
The function checks if a variable is either \code{null}, \code{NA}, or has an
assigned value. In case of vectors, the condition is set on all values.
}
