% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbraseth.R
\docType{package}
\name{jbraseth}
\alias{jbraseth}
\alias{jbraseth-package}
\title{jbraseth: jbraseth}
\description{
This package provides a collection of functions for creating violin plots, scatter plots, and more.
}
\seealso{
\code{\link{violin_plot}}, \code{\link{scatter_plot}}
}
\author{
\strong{Maintainer}: Josh Braseth \email{jbraseth@uccs.edu} (\href{https://orcid.org/0009-0004-7500-5970}{ORCID})

}
\keyword{package}
