% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{communicates with rejustify/getTableLabels API endpoint}
\usage{
labels(
  provider = NULL,
  table = NULL,
  dim = NULL,
  url = getOption("rejustify.mainUrl")
)
}
\arguments{
\item{provider}{Relevant provider (required).}

\item{table}{Relevant provider (required).}

\item{dim}{Relevant dimension from \code{provider}/\code{table} (optional).}

\item{url}{API url. By default read from global variables.}
}
\value{
list containing the pairs of value code and label
}
\description{
The function returns the list of available labels for data set \code{table}
delivered by \code{provider}. If not specified otherwise by \code{dim}, the labels are returned for all dimensions.
For the list of available provider/tables you can refer to \code{rejustify.com/repos}.
}
\examples{
#API setup
setCurl()
labels('IMF','WEO')
labels('IMF','WEO','WEO Country')

}
