% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso_model}
\alias{lasso_model}
\title{Fit a Lasso regression model using glmnet with cross-validation for lambda selection.}
\usage{
lasso_model(X, y, fold_percentage = 0.2)
}
\arguments{
\item{X}{Matrix of predictor variables. Each row represents an observation, and each column represents a predictor.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}

\item{fold_percentage}{Ratio of number of folds to sample size for cross-validation.}
}
\value{
A fitted Lasso regression model object of class "glmnet".
}
\description{
This function fits a Lasso regression model using the glmnet package with cross-validation to select the optimal lambda value. It automatically detects
the type of response variable (binary or continuous) and adjusts the family parameter in the glmnet function call accordingly.
}
\examples{
# Example usage
X <- matrix(rnorm(100 * 5), ncol = 5)
y <- sample(0:1, 100, replace = TRUE)
lasso_model(X, y, fold_percentage = 0.2)

}
