% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perte charges.R
\name{lin_losses}
\alias{lin_losses}
\title{Linear charge losses}
\usage{
lin_losses(Re, L, D, V)
}
\arguments{
\item{Re}{Reynolds number}

\item{L}{pipe length}

\item{D}{pipe diameter}

\item{V}{flow velocity}
}
\value{
Return the linear charges losses by the fluid
}
\description{
calculate the linear charges losses by a fluid flowing in a pipe
}
\examples{
lin_losses(2000,10,0.020,5)
20.38736
}
