% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{linear_model

This function is used to build multiple linear models. It can be used to carry out simple and multiple linear regression.}
\alias{linear_model

This function is used to build multiple linear models. It can be used to carry out simple and multiple linear regression.}
\alias{linear_model}
\title{Linear regression model}
\usage{
linear_model(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector of outcome values (should be a continuous variable).}
}
\value{
A list of important attributes of the fitted linear model.
}
\description{
Linear regression model
}
\examples{
data(mtcars)
linear_model(X= cbind(mtcars$cyl, mtcars$hp, mtcars$drat, mtcars$wt), Y = mtcars$mpg)

}
