% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model Function}
\usage{
linear_model(X, y)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}
}
\value{
A list containing the fitted model and predicted values.
}
\description{
This function fits a linear model using either linear regression for continuous response variables
or logistic regression for binary response variables.
}
\examples{
# Example usage:
X <- matrix(rnorm(100), ncol = 2)
y <- rnorm(100)
model <- linear_model(X, y)

}
