% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{Linearity}
\alias{Linearity}
\title{Linearity}
\usage{
Linearity(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector of outcome values (should be a continuous variable).}
}
\value{
The partial regression plots of each predictors.
}
\description{
This function is used to do the Linearity diagnosis of the linear model.
}
\examples{
data(mtcars)
Linearity(X= cbind(mtcars$cyl, mtcars$hp, mtcars$drat, mtcars$wt), Y = mtcars$mpg)

}
