% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{load_model}
\alias{load_model}
\title{Load Model Function}
\usage{
load_model(X, y, modelType)
}
\arguments{
\item{X}{Independent variables (predictors).}

\item{y}{Dependent variable (response).}

\item{modelType}{Type of machine learning model (linear, ridge, lasso, elastic net, random forest).}
}
\value{
Fitted model.
}
\description{
This function loads the appropriate model based on the specified type.
}
