% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{logIn}
\alias{logIn}
\title{Login a user}
\usage{
logIn(username, password, validFor = NULL)
}
\arguments{
\item{username}{a string representing the username}

\item{password}{a string representing the password}

\item{validFor}{an integer indicating how long the token is valid}
}
\value{
a string representing the login token
}
\description{
This function logs in a user, and returns a token the user can then use to
call restricted methods. It takes the strings username and password as input
and returns a string token as output. The user can provide an optional argument
validFor to indicate the time validity of the token.
}
