% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_merge.R
\name{multi.merge}
\alias{multi.merge}
\title{Multi-Merge Tables}
\usage{
multi.merge(tables, by, types = "inner")
}
\arguments{
\item{tables}{A list of data.tables to be merged.}

\item{by}{A character vector of column names to merge by.}

\item{types}{Type of join(s) to perform: "inner", "outer", "left", or "right".
If a single value is provided, it will be used for all merges.
If a vector is provided, it must have length equal to the number of tables minus one.}
}
\value{
A merged data.table.
}
\description{
Sequentially merge multiple tables based on a common key and join types.
}
\examples{
library(data.table)
table1 <- data.table(id = c(1, 2), value1 = c("A", "B"))
table2 <- data.table(id = c(2, 3), value2 = c("C", "D"))
table3 <- data.table(id = c(3, 4), value3 = c("E", "F"))
tables <- list(table1, table2, table3)
multi.merge(tables, by = "id", types = c("inner", "left"))

}
