% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Iterations}

\item{n}{Sample size}

\item{p}{Probability}
}
\value{
Simulates a binomial experiment and returns a table and a barplot
}
\description{
Title
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
