% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of itterations}

\item{x}{The sample}

\item{fun}{The function}

\item{alpha}{1-confidence level}

\item{cx}{Text size}

\item{...}{Other variables}
}
\value{
A barplot of the confidence interval using the bootstrap method
}
\description{
myboot2 function
}
\examples{
myboot2(iter=10000,x=c(1,2,3,4), fun="mean", alpha=1-.95,cx=1)
}
