% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Title}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Size of sample}

\item{iter}{How many times it will be repeated}

\item{p}{The probability}

\item{...}{Other parameters}
}
\value{
Histograms of mean of the sample and binomial curve
}
\description{
Title
}
\examples{
mycltb(n=4,iter=10000,p=0.3)

}
