% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{mymlbeta}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{sample}

\item{alpha}{shape 1 range}

\item{beta}{shape 2 rage}

\item{...}{Other parameters}
}
\value{
A graphical solution for shape 1 and shape 2
}
\description{
mymlbeta
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100))

}
