% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the average value}

\item{sigma}{standard deviation}

\item{a}{the interval from left critical to a}
}
\value{
A graph of a normal curve with a shaded region and the area
}
\description{
Title
}
\examples{
myncurve(mu=2,sigma=7,a=3)
}
