% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{newModelList}
\alias{newModelList}
\title{Creating ModelList class}
\usage{
newModelList(list)
}
\arguments{
\item{list}{A list of models; consider using CreateModelList generator function to create this list.}
}
\value{
A S3 class ModelList object, containing regression models
}
\description{
A function to implement class ModelList for a list input containing regression models. This class holds multiple models for manipulation.
}
\examples{
m1 <- lm(lifeExp ~ year + pop + gdpPercap, gapminder::gapminder)
m2 <- lm(lifeExp ~ year, gapminder::gapminder)
mList <- CreateModelList(m1, m2)
mList <- newModelList(mList)


}
