% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflexplore-package.R
\docType{package}
\name{nflexplore-package}
\alias{nflexplore}
\alias{nflexplore-package}
\title{nflexplore: National Football League Exploratory Data Analysis}
\description{
This R package offers essential NFL facts, game attendance, and results datasets for analysis, along with convenient functions to launch an interactive NFL statistics Shiny App and generate winning trend graphs based on team names, enhancing the NFL data exploration experience.
}
\author{
\strong{Maintainer}: Evan Ginting \email{egin0003@student.monash.edu}

}
\keyword{internal}
