% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{Normality}
\alias{Normality}
\title{Normality}
\usage{
Normality(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector of outcome values (should be a continuous variable).}
}
\value{
The histogram of the residuals.
}
\description{
This function is used to do the Normality diagnosis of the linear model.
}
\examples{
data(mtcars)
Normality(X= cbind(mtcars$cyl, mtcars$hp, mtcars$drat, mtcars$wt), Y = mtcars$mpg)

}
