% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{# Number of seats}

\item{gamma}{# Probability of a full plain}

\item{p}{# Probability of show}
}
\value{
# 2 plots and their minimum value one discrete and other continuous
}
\description{
ntickets function
}
\examples{
ntickets(400,0.02,0.95)
}
