% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{null_check}
\alias{null_check}
\title{Helper function to assign xlab and ylab to variable name if labels not supplied for plotting function}
\usage{
null_check(lab, var)
}
\arguments{
\item{lab}{The label to check.}

\item{var}{The variable name to use as the label if \code{lab} is NULL.}
}
\value{
The label to use, either \code{lab} if not NULL, or the variable name of \code{var}.
}
\description{
This function is used internally by plotting functions to assign xlab and ylab to the variable name if labels are not supplied.
}
