% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observe_landscape}
\alias{observe_landscape}
\title{observes (saves) the current landscape, can be called independantly by the user and is called by 
other observer functions relying on the landscape to be present}
\usage{
observe_landscape()
}
\description{
observes (saves) the current landscape, can be called independantly by the user and is called by 
other observer functions relying on the landscape to be present
}
