% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observe_species}
\alias{observe_species}
\title{This function can be called within the observer function to observe (save) the full species list.}
\usage{
observe_species()
}
\description{
This function can be called within the observer function to observe (save) the full species list.
}
