% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observer_plot}
\alias{observer_plot}
\title{This function can be called within the observer function to save the compressed species divergence.}
\usage{
observer_plot(plot_function = NULL)
}
\arguments{
\item{plot_function}{is the desired plot function., e.g. plot_richness, plot_species, plot_landscape}
}
\description{
This function can be called within the observer function to save the compressed species divergence.
}
\examples{
# TODO - add example
}
