% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_avg.R
\name{oil_prod}
\alias{oil_prod}
\title{Calculate the average Oil production.}
\usage{
oil_prod(x)
}
\arguments{
\item{x}{A numerical value between 1980 and 2021.}
}
\value{
A numerical value which is the average value of the total oil production
for that year if x is between 1980 and 2021 else it will return an error message
}
\description{
This function takes an input between 1980 and 2021 and returns the
average value of the oil production for the given year.
}
\examples{
oil_prod(2005)
oil_prod(1990)
oil_prod(1980)


}
