% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityDynamics.r
\name{omit}
\alias{omit}
\title{Omission of occurrences that belong to poorly sampled taxa}
\usage{
omit(dat, tax = "occurrence.genus_name", bin = "bin",
  coll = "collection_no", ref = "occurrence.reference_no", om = "ref",
  filterNA = FALSE)
}
\arguments{
\item{dat}{(data.frame): Occurrence dataset, with bin, tax and coll as column names.}

\item{tax}{(character value): The name of the taxon variable.}

\item{bin}{(character value): The name of the subsetting variable (has to be integer). For time series, this is the time-slice variable. If set to NULL, the function performs unbinned subsampling.}

\item{coll}{(character value): the variable name of the collection identifiers.}

\item{ref}{(character value): the variable name of the reference identifiers.}

\item{om}{(character value): the type of omission. "coll" omits occurrences of taxa that occurr only in one collection. "ref" omits occurrences of taxa that were described only in one reference. "binref"  will omit the set of single reference taxa that were described by more than one references, but appear in only one reference in a time bin.}

\item{filterNA}{(logical value): additional entries can be added to influence the dataset that might not have reference or collection information (NA entries). These occurrences are treated as single-collection or single-reference taxa if the na.rm argument is set to FALSE (default). Setting this argument to TRUE will keep these entries. (see example)}
}
\description{
Function to quickly omit single-collection and single-reference taxa.
}
\details{
The function returns a logical vector of the rows to be omitted. The function is embedded in the divDyn() function, but can be called independently.
}
\examples{
# omit single-reference taxa
  data(scleractinia)
  toOmit <- omit(scleractinia, bin="slc", tax="genus", om="ref")
  dat <- scleractinia[!toOmit,]

# within divDyn
  # plotting
  plotTS(stages, shading="series", boxes="per", xlim=c(260,0), 
    ylab="range-through diversity (genera)", ylim=c(0,230))
  # multiple ref/slice required
  ddNoSing <- divDyn(scleractinia, tax="genus", bin="slc", om="binref")
  lines(stages$mid, ddNoSing$divRT, lwd=2, col="red")

  # with the recent included (NA reference value)
  ddNoSingRec <- divDyn(scleractinia, tax="genus", bin="slc",
    om="binref", filterNA=TRUE)
  lines(stages$mid, ddNoSingRec$divRT, lwd=2, col="blue")
  
  # legend
  legend("topleft", legend=c("no single-ref. taxa", 
    "no single-ref. taxa,\\n with recent"), 
    col=c("red", "blue"), lwd=c(2,2))
}
