% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.delete}
\alias{path.delete}
\title{Delete a path}
\usage{
path.delete(pathId, token, recursive = FALSE)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Optional (string)}

\item{recursive}{Optional (logical)}
}
\value{
...
}
\description{
Delete a path
}
