% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.editMetaData}
\alias{path.editMetaData}
\title{Update the metadata of a path}
\usage{
path.editMetaData(
  pathId,
  token,
  description = NULL,
  attribution = NULL,
  licenseString = NULL,
  properties = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{description}{Optional (string)}

\item{attribution}{Optional (string)}

\item{properties}{Optional (object)}
}
\value{
content of the http patch request
}
\description{
Update the metadata of a path
}
