% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.editPublicAccess}
\alias{path.editPublicAccess}
\title{Update the public access of a path}
\usage{
path.editPublicAccess(pathId, token, access = NULL, hidden = NULL)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{access}{Optional (named list)}

\item{hidden}{Optional (logical)}
}
\value{
...
}
\description{
Update the public access of a path
}
