% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.favorite}
\alias{path.favorite}
\title{Add path to favorites}
\usage{
path.favorite(pathId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Add path to favorites
}
