% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_file.R
\name{path.file.add}
\alias{path.file.add}
\title{Adds a file}
\usage{
path.file.add(
  filePath,
  token,
  parentId = NULL,
  publicAccess = NULL,
  metadata = NULL
)
}
\arguments{
\item{filePath}{Mandatory (string) path to the file to upload}

\item{token}{Mandatory (string)}

\item{parentId}{Optional (uuid) id of folder to place file in}

\item{publicAccess}{Optional (named list) named list desctibing public access of the file}

\item{metadata}{Optional (named list) named list describing metadata of the file}
}
\description{
Adds a file
}
