% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_file.R
\name{path.file.download}
\alias{path.file.download}
\title{Downloads a file}
\usage{
path.file.download(pathId, filePath, token = NULL)
}
\arguments{
\item{pathId}{Mandatory (uuid) id of the folder}

\item{filePath}{Mandatory (string) location to download file to}

\item{token}{Optional (string)}
}
\description{
Downloads a file
}
