% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_folder.R
\name{path.folder.add}
\alias{path.folder.add}
\title{Adds a folder}
\usage{
path.folder.add(
  name,
  token,
  parentId = NULL,
  publicAccess = NULL,
  metadata = NULL
)
}
\arguments{
\item{name}{Mandatory (string) name for the folder}

\item{token}{Mandatory (string)}

\item{parentId}{Optional (uuid) id of the folder to place the new folder in}

\item{publicAccess}{(named list) named list describing the public access of the folder}

\item{metadata}{(named list) named list describing the metadata of the folder}
}
\description{
Adds a folder
}
