% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_folder.R
\name{path.folder.listFolder}
\alias{path.folder.listFolder}
\title{List the contents of a folder}
\usage{
path.folder.listFolder(
  pathId,
  pathTypes = NULL,
  pageStart = NULL,
  listAll = TRUE,
  includeTrashed = FALSE,
  token = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid) id of the folder}

\item{pathTypes}{Optional (list) list containing 'file', 'folder', 'raster', or 'vector'}

\item{pageStart}{Optional (uuid) from where to start the listing}

\item{listAll}{Optional (logical) default TRUE, whether to get all results or only the first page}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
List the contents of a folder
}
