% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.get}
\alias{path.get}
\title{Get information about folder}
\usage{
path.get(pathId, token = NULL)
}
\arguments{
\item{token}{a string used to authenticate the user}

\item{pathID}{a Uuid referencing a folder or the path to that folder}
}
\value{
A name value list containing information about the folder
}
\description{
This function retrieves information about a folder. The user provides the folderId
referencing the folder itself.
}
