% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_hashtag.R
\name{path.hashtag.add}
\alias{path.hashtag.add}
\title{Adds a hashtag to a map}
\usage{
path.hashtag.add(pathId, hashtag, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the map to add the hashtag to}

\item{hashtag}{Mandatory (string) the hashtag to add}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Adds a hashtag to a map
}
