% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_hashtag.R
\name{path.hashtag.delete}
\alias{path.hashtag.delete}
\title{Delete a hashtag from a given path}
\usage{
path.hashtag.delete(pathId, hashtag, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the map to remove the hashtag from}

\item{hashtag}{Mandatory (string) the hashtag to remove}

\item{token}{Mandatory (string)}
}
\description{
Delete a hashtag from a given path
}
