% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_hashtag.R
\name{path.hashtag.search}
\alias{path.hashtag.search}
\title{Retrieves all existing hashtags that contain the given string}
\usage{
path.hashtag.search()
}
\arguments{
\item{hashtag}{Mandatory (string) the string to search for}
}
\value{
...
}
\description{
Retrieves all existing hashtags that contain the given string
}
