% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_invite.R
\name{path.invite.accept}
\alias{path.invite.accept}
\title{Accept an invite to a map or a folder}
\usage{
path.invite.accept(pathId, inviteId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the map or folder}

\item{inviteId}{Mandatory (uuid) the id of the invite you wish to accept}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Accept an invite to a map or a folder
}
