% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_invite.R
\name{path.invite.getPathInvites}
\alias{path.invite.getPathInvites}
\title{Retrieve pending invites to a specific map}
\usage{
path.invite.getPathInvites(pathId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the block or folder}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Retrieve pending invites to a specific map
}
