% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_invite.R
\name{path.invite.getYourInvites}
\alias{path.invite.getYourInvites}
\title{Retrieve pending invites}
\usage{
path.invite.getYourInvites(token)
}
\arguments{
\item{token}{Mandatory (string)}
}
\description{
Retrieve pending invites
}
