% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_invite.R
\name{path.invite.revoke}
\alias{path.invite.revoke}
\title{Revoke a sent invitation}
\usage{
path.invite.revoke(pathId, inviteId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the map or folder}

\item{inviteId}{Mandatory (uuid) the id of the invite you wish to revoke}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Revoke a sent invitation
}
