% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_invite.R
\name{path.invite.send}
\alias{path.invite.send}
\title{Invites someone to a path. They will need to accept before they are added}
\usage{
path.invite.send(
  pathId,
  access,
  token,
  userId = NULL,
  email = NULL,
  sendMail = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid) the id of the block or folder to share}

\item{access}{Mandatory (object)  should be an object with properties canShare, geoFence, monthlyFee, accessLevel and processingUnits defining the access given to the user. geoFence should be an object with tiles and maxZoom. Tiles should be an array of objects with tileX, tileY and zoom.}

\item{token}{Mandatory (string)}

\item{userId}{Semi-Mandatory, at least one required (uuid) user id to send the invite to}

\item{email}{Semi-Mandatory at least one required (string) email of the user to send the invite to}

\item{sendMail}{Optional (logical) default FALSE}
}
\value{
...
}
\description{
Invites someone to a path. They will need to accept before they are added
}
