% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_member.R
\name{path.member.delete}
\alias{path.member.delete}
\title{Removes a member from a path}
\usage{
path.member.delete(pathId, userId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{userId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Removes a member from a path
}
