% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_member.R
\name{path.member.edit}
\alias{path.member.edit}
\title{Edits permissions of a user on a given path}
\usage{
path.member.edit(pathId, userId, access, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{userId}{Mandatory (uuid)}

\item{access}{Mandatory (object) Object with optional properties accessLevel, processingUnits, canShare, geoFence, with the changes in access. geoFence should be an object with tiles and maxZoom. Tiles should be an array of objects with tileX, tileY and zoom.}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Edits permissions of a user on a given path
}
