% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_member.R
\name{path.member.get}
\alias{path.member.get}
\title{Retrieves the users a path has been shared with}
\usage{
path.member.get(pathId, token = NULL, memberType = list("inherited", "direct"))
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Optional (string)}

\item{memberType}{Optional (array, vector, or list of strings) can contain "inherited" and "direct"}
}
\description{
Retrieves the users a path has been shared with
}
