% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.move}
\alias{path.move}
\title{Moves paths to a different folder}
\usage{
path.move(pathIds, parentId, token)
}
\arguments{
\item{pathIds}{Mandatory (array of uuids)}

\item{parentId}{Optional (uuid)}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
Moves paths to a different folder
}
