% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster.R
\name{path.raster.add}
\alias{path.raster.add}
\title{Add a raster}
\usage{
path.raster.add(
  name,
  token,
  parentId = NULL,
  publicAccess = NULL,
  metadata = NULL
)
}
\arguments{
\item{name}{Mandatory (string) name for the raster}

\item{token}{Mandatory (string)}

\item{parentId}{Optional (uuid) id of the folder to place the raster in}

\item{publicAccess}{Optional (named list) named list describing the raster's public access}

\item{metadata}{Optional (named list) named list describing the raster's metadata}
}
\value{
...
}
\description{
Add a raster
}
