% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster.R
\name{path.raster.edit}
\alias{path.raster.edit}
\title{Edit raster map attributes}
\usage{
path.raster.edit(
  pathId,
  token,
  interpolation = NULL,
  includesTransparent = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{interpolation}{Optional (string)}

\item{includesTransparent}{Optional (logical) default FALSE}
}
\description{
Edit raster map attributes
}
