% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster.R
\name{path.raster.editBand}
\alias{path.raster.editBand}
\title{Change the name of a band}
\usage{
path.raster.editBand(pathId, bandNumber, name, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{bandNumber}{Mandatory (integer)}

\item{name}{Mandatory (string)}

\item{token}{Mandatory (string)}
}
\description{
Change the name of a band
}
