% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_style.R
\name{path.raster.style.add}
\alias{path.raster.style.add}
\title{Adds a raster style}
\usage{
path.raster.style.add(pathId, name, method, parameters, token, default = TRUE)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{method}{Mandatory (string)}

\item{parameters}{Mandatory (object) See https://docs.ellipsis-drive.com/developers/api-v3/path-raster/styles/add-style for how to format the paramters}

\item{token}{Mandatory (string)}

\item{default}{Optional (logical) default TRUE}
}
\description{
Adds a raster style
}
