% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_style.R
\name{path.raster.style.delete}
\alias{path.raster.style.delete}
\title{Delete a raster style}
\usage{
path.raster.style.delete(pathId, styleId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{styleId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Delete a raster style
}
