% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_style.R
\name{path.raster.style.edit}
\alias{path.raster.style.edit}
\title{Edit a raster style}
\usage{
path.raster.style.edit(
  pathId,
  styleId,
  method,
  parameters,
  token,
  default = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{styleId}{Mandatory (uuid)}

\item{method}{Optional (string)}

\item{parameters}{Optional (object)}

\item{token}{Mandatory (string)}

\item{default}{Optional (logical) default = TRUE}
}
\value{
...
}
\description{
Edit a raster style
}
