% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.activate}
\alias{path.raster.timestamp.activate}
\title{Activate a timestamp}
\usage{
path.raster.timestamp.activate(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Activate a timestamp
}
