% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.add}
\alias{path.raster.timestamp.add}
\title{Add a timestamp to a raster}
\usage{
path.raster.timestamp.add(
  pathId,
  token,
  description = NULL,
  date = list(from = Sys.time(), to = Sys.time())
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{description}{Optional (string)}

\item{date}{Optional (named list) named list containing properties "to" and "from" both of type date}
}
\value{
...
}
\description{
Add a timestamp to a raster
}
