% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.analyse}
\alias{path.raster.timestamp.analyse}
\title{Request to obtain pixel values within a certain geometry}
\usage{
path.raster.timestamp.analyse(
  pathId,
  timestampIds,
  geometry,
  returnType = "all",
  approximate = TRUE,
  token = NULL,
  epsg = 4326
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampIds}{Mandatory (vector, list, or array of uuids)}

\item{geometry}{Mandatory (simple feature geometry) in WGS84}

\item{returnType}{Optional (string) either "all" or "statistics", default "all"}

\item{approximate}{Optional (logical) default TRUE}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
Request to obtain pixel values within a certain geometry
}
