% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.deactivate}
\alias{path.raster.timestamp.deactivate}
\title{Deactivate a given timestamp}
\usage{
path.raster.timestamp.deactivate(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Deactivate a given timestamp
}
