% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.edit}
\alias{path.raster.timestamp.edit}
\title{Edit a timestamp of a raster map}
\usage{
path.raster.timestamp.edit(
  pathId,
  timestampId,
  token,
  date = NULL,
  description = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{date}{Optional (named list) with properties "to" and "from" both of type date}

\item{description}{Optional (string)}
}
\description{
Edit a timestamp of a raster map
}
