% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_file.R
\name{path.raster.timestamp.file.add}
\alias{path.raster.timestamp.file.add}
\title{Upload a raster file}
\usage{
path.raster.timestamp.file.add(
  pathId,
  timestampId,
  filePath,
  token,
  fileFormat,
  epsg = NULL,
  noDataValue = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{filePath}{Mandatory (string)}

\item{token}{Mandatory (string)}

\item{fileFormat}{Optional (string)}

\item{epsg}{Optional (int)}

\item{noDataValue}{Optional (double)}
}
\value{
...
}
\description{
Upload a raster file
}
