% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_file.R
\name{path.raster.timestamp.file.delete}
\alias{path.raster.timestamp.file.delete}
\title{Delete a given upload}
\usage{
path.raster.timestamp.file.delete(pathId, timestampId, fileId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{fileId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Delete a given upload
}
