% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_file.R
\name{path.raster.timestamp.file.recover}
\alias{path.raster.timestamp.file.recover}
\title{Recover an uploaded raster file from trash}
\usage{
path.raster.timestamp.file.recover(pathId, timestampId, fileId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{fileId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Recover an uploaded raster file from trash
}
