% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_file.R
\name{path.raster.timestamp.file.trash}
\alias{path.raster.timestamp.file.trash}
\title{Move an uploaded raster file to the trash}
\usage{
path.raster.timestamp.file.trash(pathId, timestampId, fileId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{fileId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Move an uploaded raster file to the trash
}
