% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.getBounds}
\alias{path.raster.timestamp.getBounds}
\title{Request to obtain the aggregated data for a certain geometry}
\usage{
path.raster.timestamp.getBounds(pathId, timestampId, token = NULL)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\value{
simple feature dataframe (sf) containing the bounds
}
\description{
Request to obtain the aggregated data for a certain geometry
}
