% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.getRaster}
\alias{path.raster.timestamp.getRaster}
\title{Get a Raster}
\usage{
path.raster.timestamp.getRaster(
  pathId,
  timestampId,
  extent,
  style = NULL,
  threads = 1,
  token = NULL,
  showProgress = TRUE,
  epsg = 3857
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{extent}{Mandatory (named list) named list with properties xMin, xMax, yMin, yMax of type double}

\item{style}{Optional (uuid or named list describing a style ) If no style given raw data is returned. Also see https://docs.ellipsis-drive.com/developers/api-v3/path-raster/styles/add-style}

\item{token}{Optional (string)}
}
\value{
An R Raster object containing the raster
}
\description{
Get a Raster
}
