% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.getSampledRaster}
\alias{path.raster.timestamp.getSampledRaster}
\title{Get a downsampled raster}
\usage{
path.raster.timestamp.getSampledRaster(
  pathId,
  timestampId,
  extent,
  width,
  height,
  epsg = 3857,
  style = NULL,
  token = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{extent}{Mandatory (named list) a named list with properties xMin, xMax, yMin, yMax of type double}

\item{width}{Mandatory (int)}

\item{height}{Mandatory (int)}

\item{epsg}{Optional (int) default 3857 (webmercator)}

\item{style}{Optional (uuid or named list describing a style ) If no style given raw data is returned. Also see https://docs.ellipsis-drive.com/developers/api-v3/path-raster/styles/add-style}

\item{token}{Optional (string)}
}
\value{
Named list with property "raster" of type R raster object containing the downsampled raster
}
\description{
Get a downsampled raster
}
