% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_order.R
\name{path.raster.timestamp.order.add}
\alias{path.raster.timestamp.order.add}
\title{Order a download}
\usage{
path.raster.timestamp.order.add(
  pathId,
  timestampId,
  extent = NULL,
  epsg = 4326
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{extent}{Mandatory (named list) with properties xMin, xMax, yMin, yMax of type double}

\item{epsg}{Optional (int) default 4326}

\item{token}{Mandatory (string)}
}
\value{
the id of the order (uuid)
}
\description{
Order a download
}
