% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp_order.R
\name{path.raster.timestamp.order.download}
\alias{path.raster.timestamp.order.download}
\title{Download the ordered file}
\usage{
path.raster.timestamp.order.download(orderId, filePath, token)
}
\arguments{
\item{orderId}{Mandatory (uuid)}

\item{filePath}{Mandatory (string)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Download the ordered file
}
