% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.recover}
\alias{path.raster.timestamp.recover}
\title{Recover a given timestamp from the trash}
\usage{
path.raster.timestamp.recover(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Recover a given timestamp from the trash
}
