% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.trash}
\alias{path.raster.timestamp.trash}
\title{Move a given timestamp to the trash}
\usage{
path.raster.timestamp.trash(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Move a given timestamp to the trash
}
