% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.recover}
\alias{path.recover}
\title{Recover a path}
\usage{
path.recover(pathId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
Recover a path
}
