% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.rename}
\alias{path.rename}
\title{Rename a path}
\usage{
path.rename(pathId, name, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{name}{Optional (string)}

\item{token}{Optional (string)}
}
\value{
Content of the http put request
}
\description{
Rename a path
}
