% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.trash}
\alias{path.trash}
\title{Place path in trash}
\usage{
path.trash(pathId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
Place path in trash
}
