% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.unfavorite}
\alias{path.unfavorite}
\title{Removes path from favorites}
\usage{
path.unfavorite(pathId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Removes path from favorites
}
