% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_usage.R
\name{path.usage.getActiveUsers}
\alias{path.usage.getActiveUsers}
\title{Get users that used the given path}
\usage{
path.usage.getActiveUsers(pathId, token, listAl = TRUE, pageStart = NULL)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{pageStart}{Optional (uuid)}

\item{hashtag}{Mandatory (string), the hashtag to add}

\item{listAll}{Optional (logical) default TRUE}
}
\value{
...
}
\description{
Get users that used the given path
}
