% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_usage.R
\name{path.usage.getAggregatedUsage}
\alias{path.usage.getAggregatedUsage}
\title{Get aggregated usage history}
\usage{
path.usage.getAggregatedUsage(pathId, loggedIn = TRUE, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{loggedIn}{Optional (logical) default TRUE}

\item{token}{Mandatory (string)}
}
\description{
Get aggregated usage history
}
