% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_usage.R
\name{path.usage.getUsage}
\alias{path.usage.getUsage}
\title{Get usage history of a specific user for a given path}
\usage{
path.usage.getUsage(pathId, userId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{userId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Get usage history of a specific user for a given path
}
