% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector.R
\name{path.vector.add}
\alias{path.vector.add}
\title{Adds a vector}
\usage{
path.vector.add(
  name,
  token,
  parentId = NULL,
  publicAccess = NULL,
  metadata = NULL
)
}
\arguments{
\item{name}{Mandatory (string) name for the folder}

\item{token}{Mandatory (string) your token}

\item{parentId}{Optional (uuid) id of folder to place new folder in}

\item{publicAccess}{Optional (named list) list describing the public access of the folder}

\item{metadata}{Optional (named list) list describing metadata of the folder}
}
\description{
Adds a vector
}
