% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector.R
\name{path.vector.editFilter}
\alias{path.vector.editFilter}
\title{Adds indices to a vector so that you can filter on properties}
\usage{
path.vector.editFilter(pathId, propertyFilter, token)
}
\arguments{
\item{pathId}{Mandatory (uuid) id of the vector to edit}

\item{propertyFilter}{Mandatory (list) a list of named lists describing the filter}

\item{token}{Mandatory (string) your token}
}
\value{
...
}
\description{
Adds indices to a vector so that you can filter on properties
}
