% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_featureProperty.R
\name{path.vector.featureProperty.add}
\alias{path.vector.featureProperty.add}
\title{Add a feature property to a given vector}
\usage{
path.vector.featureProperty.add(
  pathId,
  name,
  featurePropertyType,
  token,
  private = FALSE,
  required = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{name}{Mandatory (string)}

\item{token}{Mandatory (string)}

\item{private}{Optional (logical) default FALSE}

\item{required}{Optional (logical) default FALSE}

\item{timestampId}{Mandatory (uuid)}

\item{featureProperty}{Mandatory (string)}
}
\description{
Add a feature property to a given vector
}
