% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_featureProperty.R
\name{path.vector.featureProperty.edit}
\alias{path.vector.featureProperty.edit}
\title{Edit the feature property for a given vector}
\usage{
path.vector.featureProperty.edit(
  pathId,
  featurePropertyId,
  token,
  private = FALSE,
  required = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{featurePropertyId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{private}{Optional (logical) default FALSE}

\item{required}{Optional (logical) default FALSE}

\item{timestampId}{Mandatory (uuid)}
}
\description{
Edit the feature property for a given vector
}
