% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_featureProperty.R
\name{path.vector.featureProperty.recover}
\alias{path.vector.featureProperty.recover}
\title{Recover a feature property for a given vector}
\usage{
path.vector.featureProperty.recover(pathId, featurePropertyId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{featurePropertyId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Recover a feature property for a given vector
}
