% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_featureProperty.R
\name{path.vector.featureProperty.trash}
\alias{path.vector.featureProperty.trash}
\title{Move a feature property for a given vector to the trash}
\usage{
path.vector.featureProperty.trash(pathId, featurePropertyId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{featurePropertyId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Move a feature property for a given vector to the trash
}
