% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_style.R
\name{path.vector.style.add}
\alias{path.vector.style.add}
\title{Add style to a vector}
\usage{
path.vector.style.add(pathId, name, method, parameters, token, default = TRUE)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{name}{Mandatory (string)}

\item{method}{Mandatory (string)}

\item{token}{Mandatory (string)}

\item{default}{Optional (logical) default TRUE}

\item{parameter}{Mandatory (object) see https://docs.ellipsis-drive.com/developers/api-v3/path-vector/styles/add-style on how to format the parameters}
}
\description{
Add style to a vector
}
