% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_style.R
\name{path.vector.style.delete}
\alias{path.vector.style.delete}
\title{Delete the style for a given vector}
\usage{
path.vector.style.delete(pathId, styleId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{styleId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Delete the style for a given vector
}
