% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_style.R
\name{path.vector.style.edit}
\alias{path.vector.style.edit}
\title{Edit the style of a vector}
\usage{
path.vector.style.edit(
  pathId,
  styleId,
  token,
  name = NULL,
  method = NULL,
  parameters = NULL,
  default = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{styleId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{method}{Optional (string)}

\item{parameters}{Optional (object)}

\item{default}{Optional (logical) default TRUE}
}
\value{
...
}
\description{
Edit the style of a vector
}
