% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.add}
\alias{path.vector.timestamp.add}
\title{Adds a vector timestamp}
\usage{
path.vector.timestamp.add(
  pathId,
  token,
  properties = NULL,
  description = NULL,
  date = list(from = Sys.time(), to = Sys.time())
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string) your token}

\item{description}{Optional (string)}

\item{date}{Optional (named list with names to and from both of type date)}
}
\value{
...
}
\description{
Adds a vector timestamp
}
