% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.deactivate}
\alias{path.vector.timestamp.deactivate}
\title{Deactivate a timestamp}
\usage{
path.vector.timestamp.deactivate(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Deactivate a timestamp
}
