% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.delete}
\alias{path.vector.timestamp.delete}
\title{Delete a vector timestamp}
\usage{
path.vector.timestamp.delete(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Delete a vector timestamp
}
