% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.edit}
\alias{path.vector.timestamp.edit}
\title{Edits a vector timestamp}
\usage{
path.vector.timestamp.edit(
  pathId,
  timestampId,
  token,
  description = NULL,
  date = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string) your token}

\item{description}{Optional (string)}

\item{date}{Optional (named list with names to and from both of type date)}
}
\value{
...
}
\description{
Edits a vector timestamp
}
