% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature.R
\name{path.vector.timestamp.feature.edit}
\alias{path.vector.timestamp.feature.edit}
\title{Edit features of a vector timestamp}
\usage{
path.vector.timestamp.feature.edit(
  pathId,
  timestampId,
  featureIds,
  features = NULL,
  token,
  showProgress = TRUE,
  levelOfDetail1 = NULL,
  levelOfDetail2 = NULL,
  levelOfDetail3 = NULL,
  levelOfDetail4 = NULL,
  levelOfDetail5 = NULL,
  cores = 1
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureIds}{Mandatory (array or list of uuids)}

\item{features}{Optional (simple features data frame (sf))}

\item{token}{Mandatory (string)}

\item{levelOfDetail1}{Optional (a list of sf geometries)}

\item{levelOfDetail2}{Optional (a list of sf geometries)}

\item{levelOfDetail3}{Optional (a list of sf geometries)}

\item{levelOfDetail4}{Optional (a list of sf geometries)}

\item{levelOfDetail5}{Optional (a list of sf geometries)}
}
\value{
...
}
\description{
Edit features of a vector timestamp
}
