% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_message.R
\name{path.vector.timestamp.feature.message.add}
\alias{path.vector.timestamp.feature.message.add}
\title{Add message to a feature}
\usage{
path.vector.timestamp.feature.message.add(
  pathId,
  timestampId,
  featureId,
  token,
  text = NULL,
  image = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{text}{Optional (string)}

\item{image}{Optional (array, matrix)}
}
\value{
...
}
\description{
Add message to a feature
}
