% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_message.R
\name{path.vector.timestamp.feature.message.get}
\alias{path.vector.timestamp.feature.message.get}
\title{Get all messages of a feature}
\usage{
path.vector.timestamp.feature.message.get(
  pathId,
  timestampId,
  featureIds = NULL,
  userId = NULL,
  messageIds = NULL,
  listAll = TRUE,
  deleted = FALSE,
  extent = NULL,
  pageStart = NULL,
  token = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{userId}{Optional (uuid)}

\item{messageIds}{Optional (list, array, or vector of uuids)}

\item{listAll}{Optional (logical) default FALSE}

\item{deleted}{Optional (logical) default FALSE}

\item{extent}{Optional (named list) with properties xMin, xMax, yMin, yMax}

\item{pageStart}{Optional (uuid)}

\item{token}{Optional (string)}
}
\value{
...
}
\description{
Get all messages of a feature
}
