% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_message.R
\name{path.vector.timestamp.feature.message.getImage}
\alias{path.vector.timestamp.feature.message.getImage}
\title{Get the image of a message}
\usage{
path.vector.timestamp.feature.message.getImage(
  pathId,
  timestampId,
  messageId,
  token = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{messageId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\value{
Depending on img_type (jpeg, png, raw bytes)
}
\description{
Get the image of a message
}
