% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_message.R
\name{path.vector.timestamp.feature.message.recover}
\alias{path.vector.timestamp.feature.message.recover}
\title{Recover message of a feature from the trash}
\usage{
path.vector.timestamp.feature.message.recover(
  pathId,
  timestampId,
  messageId,
  token
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{messageId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Recover message of a feature from the trash
}
