% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature.R
\name{path.vector.timestamp.feature.recover}
\alias{path.vector.timestamp.feature.recover}
\title{Recover a feature of a vector timestamp}
\usage{
path.vector.timestamp.feature.recover(
  pathId,
  timestampId,
  featureIds,
  token,
  showProgress = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureIds}{Mandatory (vector, list, or array of uuids)}

\item{token}{Mandatory (string)}
}
\description{
Recover a feature of a vector timestamp
}
