% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_series.R
\name{path.vector.timestamp.feature.series.add}
\alias{path.vector.timestamp.feature.series.add}
\title{Add series to a vector timestamp feature}
\usage{
path.vector.timestamp.feature.series.add(
  pathId,
  timestampId,
  featureId,
  seriesData,
  token,
  showProgress = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{seriesData}{Mandatory (data.frame)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Add series to a vector timestamp feature
}
