% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_series.R
\name{path.vector.timestamp.feature.series.info}
\alias{path.vector.timestamp.feature.series.info}
\title{Get info about series of a vector timestamp feature}
\usage{
path.vector.timestamp.feature.series.info(
  pathId,
  timestampId,
  featureId,
  token = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\description{
Get info about series of a vector timestamp feature
}
