% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_series.R
\name{path.vector.timestamp.feature.series.recover}
\alias{path.vector.timestamp.feature.series.recover}
\title{Recover series of a vector timestamp feature from the trash}
\usage{
path.vector.timestamp.feature.series.recover(
  pathId,
  timestampId,
  featureId,
  seriesIds,
  token,
  showProgress = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{seriesIds}{Mandatory (array, vector, or list of uuids)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Recover series of a vector timestamp feature from the trash
}
