% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature_series.R
\name{path.vector.timestamp.feature.series.trash}
\alias{path.vector.timestamp.feature.series.trash}
\title{Move series of a vector timestamp feature to the trash}
\usage{
path.vector.timestamp.feature.series.trash(
  pathId,
  timestampId,
  featureId,
  seriesIds,
  token,
  showProgress = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{seriesIds}{Mandatory (array, list, or vector of uuids)}

\item{token}{Mandatory (string)}
}
\description{
Move series of a vector timestamp feature to the trash
}
