% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature.R
\name{path.vector.timestamp.feature.trash}
\alias{path.vector.timestamp.feature.trash}
\title{Move a feature of a vector timestamp to the trash}
\usage{
path.vector.timestamp.feature.trash(
  pathId,
  timestampId,
  featureIds,
  token,
  showProgress = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureIds}{Mandatory (vector, list or array of uuids)}

\item{token}{Mandatory (string)}
}
\value{
...
}
\description{
Move a feature of a vector timestamp to the trash
}
