% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature.R
\name{path.vector.timestamp.feature.versions}
\alias{path.vector.timestamp.feature.versions}
\title{Get version data of a feature of a vector timestamp}
\usage{
path.vector.timestamp.feature.versions(
  pathId,
  timestampId,
  featureId,
  token = NULL,
  pageStart = NULL,
  listAll = TRUE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureId}{Mandatory (uuid)}

\item{token}{Optional (string)}

\item{pageStart}{Optional (uuid)}

\item{listAll}{Optional (logical)}
}
\value{
a named list with a result property containing the version data of the feature (simple feature data frame (sf))
}
\description{
Get version data of a feature of a vector timestamp
}
