% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_feature.R
\name{path.vector.timestamp.feature.zipLevels}
\alias{path.vector.timestamp.feature.zipLevels}
\title{Maybe use aslist (we want to combine geometries I guess...)}
\usage{
path.vector.timestamp.feature.zipLevels(
  levelOfDetail1,
  levelOfDetail2,
  levelOfDetail3,
  levelOfDetail4,
  levelOfDetail5
)
}
\description{
Maybe use aslist (we want to combine geometries I guess...)
}
