% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.getBounds}
\alias{path.vector.timestamp.getBounds}
\title{Get the bounds of a vector timestamp}
\usage{
path.vector.timestamp.getBounds(pathId, timestampId, token = NULL)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Optional (string)}

\item{timestapmId}{Mandatory (uuid)}
}
\value{
a simple feature sf object
}
\description{
Get the bounds of a vector timestamp
}
