% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.getChanges}
\alias{path.vector.timestamp.getChanges}
\title{Get the changelog of a vector timestamp}
\usage{
path.vector.timestamp.getChanges(
  pathId,
  timestampId,
  token = NULL,
  pageStart = NULL,
  listAll = FALSE
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Optional (string)}

\item{pageStart}{optional (object)}

\item{listAll}{Mandatory (logical)}

\item{actions}{Optional (object)}
}
\value{
...
}
\description{
Get the changelog of a vector timestamp
}
