% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.getFeaturesByExtent}
\alias{path.vector.timestamp.getFeaturesByExtent}
\title{Get features of a vector timestamp by the extent of the features}
\usage{
path.vector.timestamp.getFeaturesByExtent(
  pathId,
  timestampId,
  extent,
  propertyFilter = NULL,
  token = NULL,
  listAll = TRUE,
  pageStart = NULL,
  epsg = 4326,
  coordinateBuffer = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{extent}{Mandatory (named list) with names xMin xMax yMin yMax of tyoe double}

\item{propertyFilter}{Optional (object)}

\item{token}{Optional (string)}

\item{listAll}{Optional (logical) whether to list all results (default TRUE)}

\item{pageStart}{Optional (object)}
}
\value{
a simple features 'sf' object containing the features as a geometry
}
\description{
Get features of a vector timestamp by the extent of the features
}
