% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.getFeaturesByIds}
\alias{path.vector.timestamp.getFeaturesByIds}
\title{Get features of a vector timestamp by the id of the features}
\usage{
path.vector.timestamp.getFeaturesByIds(
  pathId,
  timestampId,
  featureIds,
  token = NULL,
  showProgress = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{featureIds}{Mandatory (array of uuids)}

\item{token}{Mandatory (string)}

\item{showProgress}{Mandatory (logical)}
}
\value{
a simple features 'sf' object containing the features as a geometry
}
\description{
Get features of a vector timestamp by the id of the features
}
