% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.listFeatures}
\alias{path.vector.timestamp.listFeatures}
\title{List available features of a vector timestamp}
\usage{
path.vector.timestamp.listFeatures(
  pathId,
  timestampId,
  token = NULL,
  listAll = TRUE,
  pageStart = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Optional (string)}
}
\value{
a simple features 'sf' object containing the features and the geometry. Append "properties." ("properties.id" for the featureId for instance) to the requested property
}
\description{
List available features of a vector timestamp
}
