% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_order.R
\name{path.vector.timestamp.order.add}
\alias{path.vector.timestamp.order.add}
\title{Order a download}
\usage{
path.vector.timestamp.order.add(
  pathId,
  timestampId,
  token,
  extent = NULL,
  uploadId = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{extent}{Optional (named list) with names xMin, xMax, yMin, and yMax}

\item{uploadId}{Optional (uuid)}
}
\value{
the orderId (uuid)
}
\description{
Order a download
}
