% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_order.R
\name{path.vector.timestamp.order.download}
\alias{path.vector.timestamp.order.download}
\title{Download the ordered file}
\usage{
path.vector.timestamp.order.download(orderId, filePath, token)
}
\arguments{
\item{orderId}{Mandatory (uuid)}

\item{filePath}{Mandatory (string) path to local file system where file will be saved}

\item{token}{Mandatory (string)}
}
\description{
Download the ordered file
}
