% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp_order.R
\name{path.vector.timestamp.order.get}
\alias{path.vector.timestamp.order.get}
\title{Retrieve information on all your vector downloads}
\usage{
path.vector.timestamp.order.get(token)
}
\arguments{
\item{token}{Mandatory (string)}
}
\description{
Retrieve information on all your vector downloads
}
