% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.recover}
\alias{path.vector.timestamp.recover}
\title{Recovers a vector timestamp}
\usage{
path.vector.timestamp.recover(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Recovers a vector timestamp
}
