% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_vector_timestamp.R
\name{path.vector.timestamp.trash}
\alias{path.vector.timestamp.trash}
\title{Adds a vector timestamp to the trash}
\usage{
path.vector.timestamp.trash(pathId, timestampId, token)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{timestampId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}
}
\description{
Adds a vector timestamp to the trash
}
