% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_multiple.gen3sis_space_h3}
\alias{plot_multiple.gen3sis_space_h3}
\title{Plot a set of values onto a given space}
\usage{
\method{plot_multiple}{gen3sis_space_h3}(values, space, col, no_data = NA)
}
\arguments{
\item{values}{a matrix of values with columns corresponding to sets of values, and rows corresponding to grid cells,
this will result in ncol(values) raster plots.}

\item{space}{a space to plot the values onto}

\item{col}{a vector containing a color palette. For discrete values, the first element in the vector will be assigned to zero values. If NULL, gen3sis2 internal palette will be used. Default is NULL}

\item{no_data}{what value should be used for missing data present in the values parameter}
}
\value{
no return value, called for plot
}
\description{
Plot a set of values onto a given space
}
