% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_multiple.gen3sis_space_points}
\alias{plot_multiple.gen3sis_space_points}
\title{Plot a set of values onto a given space}
\usage{
\method{plot_multiple}{gen3sis_space_points}(values, space, no_data = NA)
}
\arguments{
\item{values}{a matrix of values with columns corresponding to sets of values, and rows corresponding to grid cells,
this will result in ncol(values) raster plots.}

\item{space}{a space to plot the values onto}

\item{no_data}{what value should be used for missing data present in the values parameter}
}
\value{
no return value, called for plot
}
\description{
Plot a set of values onto a given space
}
