% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_richness}
\alias{plot_richness}
\title{Plot the richness of the given list of species on a landscape}
\usage{
plot_richness(species_list, landscape)
}
\arguments{
\item{species_list}{a list of species to use in the richness calculation}

\item{landscape}{a landscape to plot the richness onto}
}
\description{
Plot the richness of the given list of species on a landscape
}
\examples{
##TODO##

# add example
# plot_richness(species_list, landscape)
}
