% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_single.gen3sis_space_raster}
\alias{plot_single.gen3sis_space_raster}
\title{Plot a single set of values in a given space}
\usage{
\method{plot_single}{gen3sis_space_raster}(values, space, title, no_data = 0, col, legend = TRUE)
}
\arguments{
\item{values}{a named list of values, the names must correspond to cells in the space}

\item{space}{a space to plot the values onto}

\item{title}{a title string for resulting plot, the time information will be taken and appended from the space id}

\item{no_data}{what value should be used for missing values in values}

\item{col}{a vector containing a color palette. For discrete values, the first element in the vector will be assigned to zero values. If NULL, gen3sis2 internal palette will be used. Default is NULL}

\item{legend}{corresponds to the \link{raster} legend plot parameter. This can be omitted and legend is handled by raster::plot}
}
\value{
no return value, called for plot
}
\description{
Plot a single set of values in a given space
}
\examples{
\dontshow{
  #TODO change this to a working example
}

\dontrun{
# get path to output objects
datapath <- system.file(file.path("extdata", "WorldCenter"), package = "gen3sis")

# plot environmental variables at a given step
space_t_25 <- readRDS(
  file.path(datapath, "output", "config_worldcenter", "spaces", "space_t_25.rds"))
oldpar <- par(no.readonly = TRUE)
par(mfrow=c(1,2))
plot_raster_single(space_t_25$environment[,"temp"], space_t_25, "Temperature", NA)
# use col to change the color
plot_raster_single(space_t_25$environment[,"arid"], space_t_25, "Aridity", NA, 
                   col=topo.colors(5))
par(oldpar)
# note that these values were scaled by the configuration object
}
}
