% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_space}
\alias{plot_space}
\title{Plot the environment variable of a given space}
\usage{
plot_space(space)
}
\arguments{
\item{space}{the gen3sis_space to plot the environment from}
}
\value{
no return value, called for plot
}
\description{
Plot the environment variable of a given space
}
