% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_space_overview}
\alias{plot_space_overview}
\title{Plot the outline of a given space over time}
\usage{
plot_space_overview(space, env_names = NULL, breaks = NULL)
}
\arguments{
\item{space}{the input space to be plotted}

\item{env_names}{character. A vector containing variable names to plot}

\item{breaks}{numeric. A vector containing time slices to plot. If NULL, the older and the most recent time-steps will be used. Default is NULL}
}
\value{
no return value, called for plot
}
\description{
Plot the outline of a given space over time
}
