% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_species_abundance}
\alias{plot_species_abundance}
\title{Plot a species' abundance on a given space}
\usage{
plot_species_abundance(species, space, col = NULL)
}
\arguments{
\item{species}{a single species object}

\item{space}{a space object}

\item{col}{a vector containing a color palette. For discrete values, the first element in the vector will be assigned to zero values. If NULL, gen3sis2 internal palette will be used. Default is NULL}
}
\value{
no return value, called for plot
}
\description{
Plot a species' abundance on a given space
}
\examples{
\dontshow{
  #TODO change this to a working example
}

\dontrun{
  # get path to output objects
  datapath <- system.file(file.path("extdata", "WorldCenter"), package = "gen3sis")
  
  # load space and species at time step zero
  space_t_0 <- readRDS(
    file.path(datapath, "output/config_worldcenter", "spaces", "space_t_0.rds"))
  species_t_0 <- readRDS(
    file.path(datapath, "output/config_worldcenter", "species", "species_t_0.rds"))
  
  # plot species 13 range and abundances
  plot_species_abundance(species_t_0[[13]], space_t_0)
  # oh, a South American one!
  
  # plot ranges and abundances of 3 species (i.e. 1, 21 and 32)
  oldpar <- par(no.readonly = TRUE)
  par(mfrow=c(1,3))
  plot_species_abundance(species_t_0[[1]], space_t_0)
  plot_species_abundance(species_t_0[[7]], space_t_0)
  plot_species_abundance(species_t_0[[11]], space_t_0)
  par(oldpar)
  
}
}
