% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puissance.R
\name{power}
\alias{power}
\title{Pump power}
\usage{
power(Q, HMT)
}
\arguments{
\item{Q}{volumetric flow rate (m3/s)}

\item{HMT}{Overall pressure losses (m)}
}
\value{
the pomp power
}
\description{
pump power for flow in a pipe with friction losses
}
\examples{
power(1000,5,20,12)
81750

}
