% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{predict.ModelList}
\alias{predict.ModelList}
\title{Predict Method for ModelList class}
\usage{
\method{predict}{ModelList}(object, ...)
}
\arguments{
\item{object}{An object of ModelList class containing list of models; consider using \code{CreateModelList()} generator function to create this list then \code{newModelList()} to assign ModelList class.}

\item{...}{further arguments passed from generic function predict}
}
\value{
A matrix with number of columns equal to the number of models in ModelList, each column containing the predicted values for model evaluated on the new dataframe specified by user input.
}
\description{
A predict method for a ModelList class that gets all predicted values for models contained in ModelList class object.
}
\examples{
m1 <- lm(lifeExp ~ year + pop + gdpPercap, gapminder::gapminder)
m2 <- lm(lifeExp ~ year, gapminder::gapminder)
mList <- newModelList(CreateModelList(m1, m2))
data <- dplyr::filter(gapminder::gapminder, year >2000)
predict(mList, newdata= data)

}
