% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_today.R
\name{prefix_today}
\alias{prefix_today}
\title{Prefix today's date to string}
\usage{
prefix_today(x)
}
\arguments{
\item{x}{string}
}
\value{
A string in the format of "%Y%m%d x"
}
\description{
Prefix today's date in %Y%m%d format to a string
}
\examples{
prefix_today("filename.csv")
}
