% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{prepare_directories}
\alias{prepare_directories}
\title{The function prepare_directories checks if the necessary directories exist, and otherwise creates them.
This function will be called by the simulation, but is made available if the directories should be created
manually beforehand, for example to redirect the stdout to a file in the output directory.}
\usage{
prepare_directories(config_file = NA, input_directory = NA,
  output_directory = NA)
}
\arguments{
\item{config_file}{path to the config file, if NA the default config will be used}

\item{input_directory}{path to input directoy, if NA it will be derived from the config file path}

\item{output_directory}{path to output directory, if NA it will be derived from the config file path}
}
\value{
returns a named list with the paths for the input and output directories
}
\description{
The function prepare_directories checks if the necessary directories exist, and otherwise creates them.
This function will be called by the simulation, but is made available if the directories should be created
manually beforehand, for example to redirect the stdout to a file in the output directory.
}
\examples{
##TODO##

# add example

# prepare_directories(...)
}
