% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_screening.R
\name{pScore}
\alias{pScore}
\title{Calculate p-value for Univariate Association Test}
\usage{
pScore(x, y)
}
\arguments{
\item{x}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}
}
\value{
The p-value indicating the significance of association between x and y.
}
\description{
This function calculates the p-value for a univariate association test between two variables x and y.
The appropriate statistical test is selected based on the data types of x and y.
If both x and y are categorical, chi-squared test is performed.
If x is numeric and y is categorical, ANOVA or Fisher's Exact Test is conducted.
If x is categorical and y is numeric, Kruskal-Wallis test or Simple Linear Regression is used.
If both x and y are numeric, Pearson's correlation test is applied.
}
