% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{pseudorecords}
\alias{pseudorecords}
\title{Create pseudo-records from an FAD-LAD matrix}
\usage{
pseudorecords(x, tax = "row.names", fad = "FAD", lad = "LAD", binned = TRUE)
}
\arguments{
\item{x}{FAD-LAD matrix or data.frame.}

\item{fad}{The column name of the first appearances.}

\item{lad}{The column name of the last appearances.}

\item{binned}{Logical flag indicating whether the FAD and LAD dates are given as integer bins or not (only binned).}
}
\value{
An occurrence record data.frame
}
\description{
Create a fully-sampled occurrence record data frame from an FAD-LAD matrix. 
Defaulted to be used with the fadlad function. Can also be used to work with \code{equilibrialBD}.
}
