% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{raster_plot}
\alias{raster_plot}
\title{plots continuous data in raster cells
NOTE: expects config.yml in same or parent folder with x_sim, y_sim, x_obs, y_obs}
\usage{
raster_plot(df, label = NA, metric = NA, dir = 1, ColPalette = "Blues")
}
\arguments{
\item{df}{dataframe with columns x, y, z, with x & y being coordinates and z the value to be plotted}

\item{label}{character vector that should be shown as a label of the plot (also refered to as tag, e.g. A)}

\item{metric}{character vector that should be shown as the label in the legend of values z (e.g. richness)}

\item{dir}{integer indicating the direction of the colour vector (can be 1 or -1)}

\item{ColPalette}{colourBrewer palette that should be used for the colours of the raster plot}
}
\value{
ggplot with coloured raster (input for plot_SimulationStep function)
}
\description{
plots continuous data in raster cells
NOTE: expects config.yml in same or parent folder with x_sim, y_sim, x_obs, y_obs
}
