% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSplit.R
\name{ratesplit}
\alias{ratesplit}
\title{Test of rate split (selectivity testing)}
\usage{
ratesplit(dat, sel, tax = "genus", bin = "slc", rate = "pc",
  method = "AIC", AICc = T, na.rm = T, alpha = NULL,
  output = "simple")
}
\arguments{
\item{dat}{(data.frame): the data frame containing PBDB occurrences.}

\item{sel}{(character): Variable name to do the splitting of the dataset. Can have only two levels.}

\item{tax}{(char): variable  name of the occurring taxa (variable type: factor) - such as "occurrence.genus_name"}

\item{bin}{(char): variable name of the time slice numbers of the particular occurrences (variable type: int)- such as "slc" or whatever. Bin numbers should be in ascending order,can contain NA's, it can start from a number other than 1 and must not start with 0.}

\item{rate}{(character): The rate metric. Currently only the per capita rates of Foote (2000) are available.}

\item{method}{(character): Either "AIC", "binom" or "combine". The "AIC" method calculates the Akaike weights of the single and dual rate models. The "binom" method assumes a binomial error distribution of the counts that are necessary for the rate calculations. The "combine" method shows slices that pass both tests, the "AIC" being usually the stronger.}

\item{AICc}{(logical): Only applicable for the "AIC" method. Toggles whether the small sample corrected AIC (AICc) should be used instead of the regular one.}

\item{alpha}{(num): Threshold for the between meaningful and meaningless split. If method=="AIC", the value corresponds to the minimum weight value the dual model should have , by default it is 0.89, which corresponds to the likelihood ratio of 8. If method=="binom", the value corresponds to the alpha value of the binomial test. If method=="combine" than two alpha values are required (1st for the AIC, 2nd for the binomial test). If alpha is NULL, than the default values will be used.}

\item{output}{(character): Either "simple" or "full". Simple returns the indices of the series where selectivity can be suggested. "full" returns a matrix of Akaike weights, or binomial probabilities.}
}
\description{
This function will determine whether there are meaningful differences between the taxonomic rates in the individual time slices of two subsets of an occurrence database.
}
\details{
Splitting an occurrence database to its subsets secreases the amount of information passed to the rate calculations and therefore the precision of the individual estimates. In order to assess the subsets individually and compare them, it is advised to test whether they the split is meaningful, given the total data. Examples of this use can be found in Kiessling and Simpson (2011), Global Change Biology 17, 56-67 and Kiessling and Kocsis (2015), Paleobiology 41, 402-414.

The meaningfulness of the split is dependent on the estimate accurracy and the magnitude of the difference.
}
\examples{
# example with the coral dataset of Kiessling and Kocsis (2015)
data(corals)
data(stages)

# split by ecology
  z<-corals[corals$ecology=="z",]
  az<-corals[corals$ecology=="az",]

# calculate diversity dynamics
ddZ<-divDyn(z, tax="genus", bin="slc")
ddAZ<-divDyn(az, tax="genus", bin="slc")

# origination rate plot
plotTS(stages, boxes="per", shading="series", xlim=54:95, 
  ylab="raw per capita originations")
lines(stages$mid, ddZ$oriPC, lwd=2, lty=1, col="blue")
lines(stages$mid, ddAZ$oriPC, lwd=2, lty=2, col="red")
legend("topright", inset=c(0.1,0.1), legend=c("z", "az"), 
  lwd=2, lty=c(1,2), col=c("blue", "red"), bg="white")

# The ratesplit function
rs<-ratesplit(rbind(z, az), sel="ecology", tax="genus", bin="slc")
rs

# display selectivity with points
# select the higher rates
selIntervals<-cbind(ddZ$oriPC[rs$ori], ddAZ$oriPC[rs$ori])
groupSelector<-apply(selIntervals, 1, function(x) x[1]<x[2])
# draw the points
points(stages$mid[rs$ori[groupSelector]], ddAZ$oriPC[rs$ori[groupSelector]],
  pch=16, col="red", cex=2)
points(stages$mid[rs$ori[!groupSelector]], ddZ$oriPC[rs$ori[!groupSelector]],
  pch=16, col="blue", cex=2)


}
