% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reynolds number.R
\name{Re}
\alias{Re}
\title{Reynolds number calculation}
\usage{
Re(rho, V, D, nu)
}
\arguments{
\item{rho}{density of the fluid (kg/m3)}

\item{V}{flow velocity (m/s)}

\item{D}{Pipe diameter (m)}

\item{nu}{dynamic viscosity (Pa.s)}
}
\value{
The value of the Reynolds number
}
\description{
Calculate the Reynolds number for a fluid flowing in a pipe
and determine its flow regime.
}
\examples{

Re(1000,2,0.1,0.001)
Turbulent flow
20000
}
