% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reynolds number 2.R
\name{Re_Q}
\alias{Re_Q}
\title{Reynolds number calculation}
\usage{
Re_Q(rho, Q, nu, D)
}
\arguments{
\item{rho}{density of the fluid (kg/m3)}

\item{Q}{volumetric flow rate (m3/s)}

\item{nu}{dynamic viscosity (Pa.s)}

\item{D}{pipe diameter (m)}
}
\value{
The value of the Reynolds number
}
\description{
Calculate the Reynolds number for a fluid flowing in a pipe
using volumetric flow rate Q.
}
\examples{

Re(1000,2,0.1,0.001)
Turbulent flow
20000
}
