% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abscensustablebuilder.R
\name{read_abscensustablebuilder}
\alias{read_abscensustablebuilder}
\title{Extracts and formats ABS Census TableBuilder output}
\usage{
read_abscensustablebuilder(table, totals = FALSE, short_names = FALSE)
}
\arguments{
\item{table}{ABS Census TableBuilder .xlsx export (.csv exports will not work)}

\item{totals}{Default FALSE. TRUE will include the total columns/rows in the export which can be annoying.}

\item{short_names}{Default FALSE. TRUE will label columns using TableBuilder abbreviations instead of the entire label e.g. 'AGE5P' instead of 'Age in Five Year Groups'.}
}
\value{
A data frame with up to 4 columns: 'row', 'column', 'wafer' columns depending on ABS variables and a 'value' variable.
}
\description{
Takes ABS Census TableBuilder .xlsx files and reorganises rows, columns, and wafers into a 'long' formatted dataframe.
}
