% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read the raw data set into a data frame}
\usage{
read_dataset(file)
}
\arguments{
\item{file}{data set file}
}
\value{
data frame
}
\description{
Read the raw data set into a data frame
}
\examples{

}
