% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_excel.R
\name{read_dataset_excel}
\alias{read_dataset_excel}
\title{Read data set from excel files}
\usage{
read_dataset_excel(
 
    file_response = "../SlimStampen_data_examples/vocatrainer/nmd_live_vocatrainer_public_response.xlsx",
 
    file_lesson = "../SlimStampen_data_examples/vocatrainer/nmd_live_vocatrainer_public_lesson.xlsx",
 
    file_fact = "../SlimStampen_data_examples/vocatrainer/nmd_live_vocatrainer_public_fact.xlsx"
)
}
\arguments{
\item{file}{Excel files. If no file is provided the example .xlsx file will
be loaded.}
}
\value{
data frame
}
\description{
\code{read_dataset_excel} reads the raw data set from multiple excel files
into a data frame with the correct settings
}
