% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{sets the token and email as global variables}
\usage{
register(token = NULL, email = NULL)
}
\arguments{
\item{token}{API token.}

\item{email}{E-mail address for the account.}
}
\value{
errors only
}
\description{
The function stores the account details into memory to be easier accessed by
rejustify package. The email must correspond to the token that was assigned to it. To register an
account visit \code{rejustify.com}.
}
\examples{
register(token = "YOUR_TOKEN", email = "YOUR_EMAIL")

}
