% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_data2.R
\docType{data}
\name{resource_data2}
\alias{resource_data2}
\title{Resource data}
\format{
This data frame has 55422 rows and 8 variables.
\describe{
\item{Entity}{Name of the country}
\item{Year}{Year}
\item{Gas reserves}{Gas reserves of the country}
\item{Oil reserves}{Oil reserves of the country}
\item{Coal reserves}{Coal reserves of the country}
\item{Population}{Population of the country}
\item{type}{Type of consumption or production}
\item{quantity}{quantity of the type}
}
}
\source{
\url{https://ourworldindata.org/explorers/natural-resources}
}
\usage{
resource_data2
}
\description{
This dataset contains information about fossil fuel production, consumption,
population of the country,year and the total reserves of natural resources
of different countries from 1973 to 2021.
}
\keyword{datasets}
