% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{rf_model}
\alias{rf_model}
\title{Fit a Random Forest model using the randomForest package.}
\usage{
rf_model(X, y)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}
}
\value{
A fitted Random Forest model object of class "randomForest".
}
\description{
Fit a Random Forest model using the randomForest package.
}
\examples{
# Example usage for classification
X <- matrix(rnorm(100 * 2), ncol = 2)
y <- sample(c(0, 1), 100, replace = TRUE)
rf_model <- rf_model(X, y)
# Example usage for regression
X <- matrix(rnorm(100 * 3), ncol = 3)
y <- rnorm(100)
rf_model <- rf_model(X, y)

}
