% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_model}
\alias{ridge_model}
\title{Ridge Regression}
\usage{
ridge_model(X, y)
}
\arguments{
\item{X}{matrix of the predictor variables}

\item{y}{matrix of the output variables}
}
\value{
coefficient matrix of the covariates
}
\description{
Ridge Regression
}
