% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz77enc.R
\name{RLE_compressor}
\alias{RLE_compressor}
\title{RLE compressor. Encodes single-element character vector or vector of characters}
\usage{
RLE_compressor(s, as.bytes = F, to.raw = F, as.vec = T)
}
\arguments{
\item{s}{character of vector of characters}

\item{as.bytes}{Logical value, default FALSE. If T - the counts of characters are
converted to bytes, if F - are left as they are. Is useful for
text with digits or with long substrings (e.g 'aaa123aaa' or 'asddddddddddda')}

\item{to.raw}{Logical value, default FALSE. If T - returns raw-data, else - characters vector}

\item{as.vec}{Logical value, default TRUE. If T - returns vector of characters,
else - one-element character vector}
}
\value{
Raw-data or vector of characters of one-element character
}
\description{
RLE compressor. Encodes single-element character vector or vector of characters
}
\examples{
test_rle <- c("aaadf1111111111")
 RLE_compressor(test_rle, as.vec=F)
 RLE_compressor(test_rle, as.bytes=T)
 RLE_compressor(test_rle, to.raw = T)
}
