% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz77enc.R
\name{RLE_compressor}
\alias{RLE_compressor}
\title{RLE decoder. Decodes output from the encoder}
\usage{
RLE_compressor(s, as.bytes = F, from.raw = F, as.vec = T)
}
\arguments{
\item{s}{output from the encoder (raw, character or vector)}

\item{as.bytes}{Logical value, default FALSE. If T - the counts of characters are
converted from bytes, if F - are left as they are. Should be the same
as for the encoder}

\item{from.raw}{Logical value, default FALSE. If T - decodes raw-data, else - characters vector
Should be the same as for the encoder}

\item{as.vec}{Logical value, default TRUE. If T - returns vector of characters,
else - one-element character vector}
}
\value{
Vector of characters of one-element character
}
\description{
RLE compressor. Decodes output from the encoder
}
\examples{
test_rle <- c("aaadf1111111111")
 RLE_decompressor(RLE_compressor(test_rle, as.bytes=T), as.bytes=T)
 RLE_decompressor(RLE_compressor(test_rle, to.raw = T), from.raw=T)
}
