% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{run_app}
\alias{run_app}
\title{Function to launch shiny app}
\usage{
run_app()
}
\value{
This function returns a shiny app created using the data loaded in the
package.
}
\description{
This function launches the shiny app.
}
