% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ausbushfire_app}
\alias{run_ausbushfire_app}
\title{Launch the ausbushfire Shiny app}
\usage{
run_ausbushfire_app()
}
\value{
Invisibly returns the app object (called for side effects).
}
\description{
Starts the Shiny application bundled with the package (located under
\code{inst/shiny}). This is the recommended way for users to run the app after
installing the package.
}
