% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampstat.R
\name{sampstat}
\alias{sampstat}
\title{Basic sampling summary for every bin}
\usage{
sampstat(dat, tax = "occurrence.genus_name", bin = "slc",
  coll = "collection_no", ref = "occurrence.reference_no",
  noNAstart = FALSE)
}
\arguments{
\item{dat}{(data.frame): the occurrence dataset.}

\item{tax}{(char): the column name of taxon names.}

\item{bin}{(char): the column name of bin names.}

\item{coll}{(char): the column name of collection numbers.}

\item{ref}{(char): the column name of reference numbers.}
}
\description{
This function will return the basic sampling summaries of a dataset
}
\details{
A number of sampling related variables are calculated with the function.

occs: The number of occurrences in each time bin.

colls: The number of collections in each time bin.

refs: The number of references in each time bin.

SIBs: The number of Sampled-In-Bin taxa in each time bin.

occ1: The number of taxa in each time bin, that occur in only 1 collection.

ref1: The number of taxa in each time bin, that occur in only 1 reference.

occ2: The number of taxa in each time bin, that occur in exactly 2 collections.

ref2: The number of taxa in each time bin, that occur in exactly 2 references.

dom: Dominance, the proportion of occurrences in the time bin that belong to the most frequent taxon.

u: Good's u, coverage estimator based on the number of single-collection taxa (occ1).

uPrime: Good's u, coverage estimator based on the number of single-reference taxa (ref1).

chao1occ: Chao1 extrapolation estimator, based on the the number of single-collection and two-collection taxa (occ1).

chao1ref: Chao1 extrapolation estimator, based on the the number of single-reference and two-reference taxa (occ2).
}
\examples{
data(scleractinia)
# slice-specific sampling
basic<-sampstat(scleractinia, tax="genus", bin="slc")

# subsampling diagnostic
 subStats<-subsample(scleractinia, method="cr", tax="genus", FUN=sampstat, bin="slc", q=100,noNAstart=F)

}
