% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_abund_matrix}
\alias{save_abund_matrix}
\title{This function can be called within the observer function to save a presence-absence matrix for each species.
saves matrix for each timestep in respective "abundance_matrices" folder within output folder
matrix format: first two columns x and y contain x&y coordinates, subsequent columns contain abundance values for each cell (rows) for each species (cols)}
\usage{
save_abund_matrix()
}
\value{
abundance matrix
}
\description{
This function can be called within the observer function to save a presence-absence matrix for each species.
saves matrix for each timestep in respective "abundance_matrices" folder within output folder
matrix format: first two columns x and y contain x&y coordinates, subsequent columns contain abundance values for each cell (rows) for each species (cols)
}
