% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_abundance}
\alias{save_abundance}
\title{This function can be called within the observer function to save the species abundances.}
\usage{
save_abundance()
}
\description{
This function can be called within the observer function to save the species abundances.
}
\examples{

}
