% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_divergence}
\alias{save_divergence}
\title{This function can be called within the observer function to save the compressed species divergence.}
\usage{
save_divergence()
}
\description{
This function can be called within the observer function to save the compressed species divergence.
}
\examples{

}
