% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_landscape}
\alias{save_landscape}
\title{This function can be called within the observer function to save 
the current landscape, can be called independently by the user and is called by 
other observer functions relying on the landscape to be present (e.g. save_species)}
\usage{
save_landscape()
}
\value{
no return value, called for side effects
}
\description{
This function can be called within the observer function to save 
the current landscape, can be called independently by the user and is called by 
other observer functions relying on the landscape to be present (e.g. save_species)
}
\examples{
\dontrun{
  ## save landscape from within observer for each species
  # this functions should be called inside the end_of_timestep_observer function at the config file:
  save_landscape()
}
}
\seealso{
\code{\link{save_species}}
}
