% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_occupancy}
\alias{save_occupancy}
\title{This function can be called within the observer function to save the current occupancy pattern}
\usage{
save_occupancy()
}
\description{
This function can be called within the observer function to save the current occupancy pattern
}
