% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_phylogeny}
\alias{save_phylogeny}
\title{This function can be called within the observer function to save the current phylogeny.}
\usage{
save_phylogeny()
}
\description{
This function can be called within the observer function to save the current phylogeny.
}
\examples{
##TODO##
}
