% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_species}
\alias{save_species}
\title{This function can be called within the observer function to save the full species list.}
\usage{
save_species()
}
\description{
This function can be called within the observer function to save the full species list.
}
