% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_traits}
\alias{save_traits}
\title{This function can be called within the observer function to save the species traits.}
\usage{
save_traits()
}
\value{
no return value, called for side effects
}
\description{
This function can be called within the observer function to save the species traits.
}
\examples{
\dontrun{
  ## save the current traits pattern from within observer for each population of each species
  # this functions should be called inside the end_of_timestep_observer function at the config file:
  save_traits()
}
}
\seealso{
\code{\link{save_species}}
}
