% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_abs.R
\name{select_abs}
\alias{select_abs}
\title{Select relevant columns from readABS data}
\usage{
select_abs(read_abs_output)
}
\arguments{
\item{read_abs_output}{output of readabs::read_abs function.}
}
\value{
A tibble of read_abs output with only 3 columns: 'date', 'series', and 'value'.
}
\description{
Removes all columns from ABS data aside from 'date', 'series', and 'value'. It is useful when the other information provided in the output, e.g. table_no, sheet_no, table_title, frequency, unit, etc, is not required such as when downloading a specific series or table and you know what it contains. Alternatively you're just lazy and don't want to type out the code yourself every time - much like myself when creating this function.
}
