% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepseries_abs.R
\name{sepseries_abs}
\alias{sepseries_abs}
\title{Separate 'series' column into separate columns}
\usage{
sepseries_abs(read_abs_output, lastcol = FALSE)
}
\arguments{
\item{read_abs_output}{output of readabs::read_abs function - ideally after justinduytran::select_abs.}

\item{lastcol}{boolean to include the last column or not. Default FALSE. TRUE includes the last column which is usually just a space given the ABS likes to end their series descriptions with "; " and is not recommended.}
}
\value{
A tibble of read_abs output but with the 'series' column separated into component columns.
}
\description{
Separates the 'series' column in ABS data into it's constituent parts by ';'. Useful when importing a table intending to use several series and multiple useful descriptors are in the series field i.e. "Estimated Resident Population ;  |SEX| ;  |AGE| ;". One day I'll build in a column renaming argument.
}
