% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisaillement.R
\name{Shear_stress}
\alias{Shear_stress}
\title{Shear stress}
\usage{
Shear_stress(Re, L, D, V, rho)
}
\arguments{
\item{Re}{Reynolds number}

\item{L}{pipe length}

\item{D}{pipe diameter}

\item{V}{fluid velocity}

\item{rho}{density}
}
\value{
the shear stress value
}
\description{
Determine Shear stress(en pa) of a given
fluid on the walls of a conduit.
}
\examples{

Shear_stress(2500,25,0.01,10,1000)
1280
}
