% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ensemble.R
\name{simple_ensemble}
\alias{simple_ensemble}
\title{Simple Ensemble Learning}
\usage{
simple_ensemble(
  X,
  y,
  models = c("elastic_net"),
  r.bagging = NULL,
  is.ensemble = FALSE,
  k = NULL
)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. Can be continuous and normally distributed or binary responses.}

\item{models}{A character vector specifying the types of models to include in the ensemble. Options include "elastic_net"
and "random_forest". They are the default option too.}

\item{r.bagging}{Number of times model will perform sampling with replacement for the samples. Default is 50.}

\item{is.ensemble}{Logical indicating whether to perform ensemble learning of models passed as parameter.}

\item{k}{Top number of predictors to select in the pre-screening step. Default is NULL.}
}
\value{
A list containing the results of the combined ensemble learning process(That is, with or without bagging,
with or without the pre screening for Top K predictors, with or without combined predictions of models)
}
\description{
This function performs ensemble learning using a combination of linear models (such as Elastic Net) and tree-based models (such as Random Forest).
It conducts preliminary data screening, trains multiple models, and optionally performs ensemble learning.
}
\examples{
# Example usage:
set.seed(123)
n <- 1000
p <- 10
X <- matrix(rnorm(n * p), ncol = p)
y <- sample(0:1, n, replace = TRUE)

# Call simple_ensemble function with Elastic Net model with R parameter for bagging
results <- simple_ensemble(X, y, models = c("elastic_net"), r.bagging = 50, is.ensemble = FALSE, k = NULL)
}
