% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{equilibrialBD}
\alias{equilibrialBD}
\title{Equilibrial Origination-Extinction (Birth-Death) dynamics with Mass Extinctions}
\usage{
equilibrialBD(
  S = 200,
  K = 200,
  pExtBase = 0.001,
  fExt = expExt,
  pOriBase = 0.001,
  fOri = expOri,
  steps = 22400,
  meP = 0.05,
  me = 9979:9999,
  plot = TRUE,
  counter = TRUE,
  extNA = TRUE
)
}
\arguments{
\item{S}{The number of starting shites}

\item{K}{Carrying capacity. A single value or a vector of values.
If this is a single value, then it is replicated 'steps' times.}

\item{pExtBase}{Baseline per timestep, per lineage probability of extinction.}

\item{fExt}{Function to scale extinction probabilities to achieve equilibrial dynamics.
The function has to have the arguments: the baseline probability (p), the equilibrial richness (k),
and standing richness (s) in the previous time step.}

\item{pOriBase}{Baseline per timestep, per lineage probability of origination.}

\item{steps}{The number of time steps.}

\item{meP}{The extinction probability during a mass extinctions.}

\item{me}{Mass extinction intervals (time step numbers).}

\item{plot}{Should a simple plot of the carrying capacity and the simulated richness be plotted?}

\item{extNA}{Should the starting and extant taxa be denoted with the LAD of NA (TRUE), or with the last time step?}
}
\value{
A data.frame with four columns: id, fad, lad, and parent. See description for
for the details.
}
\description{
Extinction and origination cannot happen in the same timestep. The returned data.frame
records lineages as rows, and it has four columns:
- id: the unique identifier of the lineage
- fad: the first appearance date of the lineage (time step number).
- lad: the last appearance date of the lineage (time step number). The extant lineages
have missing values as their LAD
- parent: the parent (ancestor) linage id.
The model was originally described in Kocsis, 2015. Analysis of global diversity patterns and dynamics of selected Mesozoic marine invertebrate groups (ELTE, PhD Thesis)
}
\examples{
set.seed(64)
K<- wienerts(S=1000, sd=20)
plot(K)
one <- equilibrialBD(S=1000, K=K, pExt=0.001, pOri=0.001)
}
