% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{wienerts}
\alias{wienerts}
\alias{expOri}
\alias{expExt}
\title{Tools useful for simulation dodels}
\usage{
wienerts(S = 1000, steps = 22400, sd = 0.5, rounding = TRUE)

expOri(p, k, s)

expExt(p, k, s)
}
\arguments{
\item{S}{Starting value}

\item{steps}{Number of steps to be generated}

\item{sd}{Standard deviation of a Gaussian distribution to be taken in every time step, will be rounded to integer values.}

\item{rounding}{Logical - should integer series be produced?}

\item{p}{The base probabilities.}

\item{k}{Carrying capacity.}

\item{s}{Richness in previous slice}
}
\value{
A numeric vector (time series) with steps as the the length, starting with S.
}
\description{
Helpful collection of functions.
}
\details{
The function \code{wienerts} generate a time series of integer values with a random walk (Wiener process). \code{expOri} and \code{expExt} are
used for scaling origination and extinction probabilities based on a carrying capacity to reach equilibrial dynamics.
}
\examples{
K<- wienerts(S=500, sd=10)
plot(K)
}
