% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuPower_norm.R
\name{simuPower_norm}
\alias{simuPower_norm}
\title{Simulation-based circadian power calculation assuming Normal errors in the cosiner model.}
\usage{
simuPower_norm(
  B_MC = 10000,
  n,
  A,
  sigma,
  phi = 0,
  period = 24,
  M = 10,
  rho = 0,
  GmoduleSize = 50,
  cts = NULL,
  alpha = 0.05
)
}
\arguments{
\item{B_MC}{Numer of Monte-Carlo simulations. Default is 10000.}

\item{n}{Sample size.}

\item{A}{Amplitude of the sinusoidal curve: \eqn{A * sin(2\pi/period * (phi + cts)) + M}.}

\item{sigma}{Standard deviation \eqn{\sigma} of each gene in the Normal error term \eqn{N(0,\Sigma)},
where \eqn{\Sigma} is a symmetric matrix of size GmoduleSize with diagonal elements being \eqn{\sigma^2} and off-diagonal
elements being \eqn{\sigma^2\rho}.}

\item{phi}{Phase shift of the sinusoidal curve. Default is 0.}

\item{period}{Period of the sinusoidal curve. Default is 24.}

\item{M}{MESOR (Midline Statistic Of Rhythm, a rhythm-adjusted mean) of the sinusoidal curve. Default is 10.}

\item{rho}{Correlation coefficient in the Normal error term \eqn{N(0,\Sigma)},
where \eqn{\Sigma} is a symmetric matrix of size GmoduleSize with diagonal elements being \eqn{\sigma^2} and off-diagonal
elements being \eqn{\sigma^2\rho}.}

\item{GmoduleSize}{Correlated genes are simulated in modules of size GmoduleSize. Only used when \eqn{\rho} is not 0.}

\item{cts}{Circadian times of samples.}

\item{alpha}{Type I error control. Default is 0.05.}
}
\value{
A vector of empirical type I error and power.
}
\description{
Simulation-based circadian power calculation assuming Normal errors in the cosiner model.
}
\details{
Calculate power of simulated data with independent Normal errors in the cosiner model.
}
\examples{
B_MC = 10000
n = 100
cts = seq(0,24,length.out = n+1)[-1]
A = 1
sigma = 1
phi = 0
M = 10
simuPower_norm(B_MC, n, A, sigma, phi=0, period = 24, cts=cts, alpha = 0.05)
}
\author{
Wei Zong, Zhiguang Huo
}
