% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuPower_outlier.R
\name{simuPower_outlier}
\alias{simuPower_outlier}
\title{Simulation-based circadian power calculation assuming independent Normal errors with outliers in the cosiner model.}
\usage{
simuPower_outlier(
  B_MC = 10000,
  n,
  A,
  sigma,
  outlier.perc = 0.05,
  phi = 0,
  period = 24,
  M = 10,
  cts = NULL,
  alpha = 0.05
)
}
\arguments{
\item{B_MC}{Numer of Monte-Carlo simulations. Default is 10000.}

\item{n}{Sample size.}

\item{A}{Amplitude of the sinusoidal curve: \eqn{A * sin(2\pi/period * (phi + cts)) + M}.}

\item{sigma}{\eqn{\sigma} in the independent Normal error term \eqn{N(0,\sigma)}.}

\item{outlier.perc}{The percentage of outlying samples. Default is 0.05.}

\item{phi}{Phase shift of the sinusoidal curve. Default is 0.}

\item{period}{Period of the sinusoidal curve. Default is 24.}

\item{M}{MESOR (Midline Statistic Of Rhythm, a rhythm-adjusted mean) of the sinusoidal curve. Default is 10.}

\item{cts}{Circadian times of samples.}

\item{alpha}{Type I error control. Default is 0.05.}
}
\value{
A vector of empirical type I error and power.
}
\description{
Simulation-based circadian power calculation assuming independent Normal errors with outliers in the cosiner model.
}
\details{
Calculate power of simulated data assuming independent Normal errors with outliers in the cosiner model.
}
\examples{
B_MC = 10000
n = 100
cts = seq(0,24,length.out = n+1)[-1]
A = 1
sigma = 1
phi = 0
M = 10
outlier.perc = 0.05
simuPower_outlier(B_MC, n, A, sigma, outlier.perc, phi, period = 24, M=10, cts,
alpha = 0.05)
}
\author{
Wei Zong, Zhiguang Huo
}
