% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perte singuliere.R
\name{sing_losses}
\alias{sing_losses}
\title{Singular pressure losses}
\usage{
sing_losses(K, V, rho)
}
\arguments{
\item{K}{singular loss coefficient}

\item{V}{flow velocity in  the pipe (m/s)}

\item{rho}{density of the fluid (kg/m3)}
}
\value{
Singular pressure losses
}
\description{
Singular head losses in pipe fittings
}
\examples{

sing_losses(0.4,10,1000)
196.2

}
