% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_config.R
\name{skeleton_config}
\alias{skeleton_config}
\title{, paste0("# Date: ", format(Sys.Date(), format="%d.%m.%Y")),}
\usage{
skeleton_config()
}
\description{
, paste0("# Date: ", format(Sys.Date(), format="%d.%m.%Y")),
}
