% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimStampeRData.R
\docType{package}
\name{SlimStampeRData}
\alias{SlimStampeRData}
\title{SlimStampeRData: Analyzing SlimStampen App Data}
\description{
The SlimStampeRData package provides three categories of important functions:
basic data descriptions, plotting column information and plotting calculated
information.
}
\section{Basic data descriptions}{

These functions...
}

