% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityDynamics.r
\name{spCleanse}
\alias{spCleanse}
\title{Cleanse Species Vector}
\usage{
spCleanse(vec, mode = "simple", collapse = "_")
}
\arguments{
\item{vec}{Character vector, the vector containing species names with qualifiers of open taxonomy.}

\item{mode}{Character value, either "simple" or "debug". "simple" will return the cleaned species name vector, "debug" returns a data table that allows one by one checking.}

\item{collapse}{Character value, this argument will be passed to the paste function's argument of the same name. The character value to be inserted between the genus and species names.}
}
\description{
This function will get rid of
}
\details{
This version will not keep subgenera, and will assign species to the base genus. The following qualifiers will be omitted:
"n.", "sp.", "?", "gen.", "aff.", "cf.", "ex", "gr.", "subgen.", "spp" and informal species designated with letters. Entries with "informal" and "indet." in them will also be invalidated.
}
\examples{
examp <- c("Genus cf. species", "Genus spp.", "Family indet.", "Mygenus yourspecies", "Okgenus ? questionsp")
spCleanse(examp) 
}

