% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tens}
\alias{tens}
\title{The 10 million year resolution timescale of the Paleobiology Database}
\format{
A \code{data.frame} with 49 observations and 9 variables:
	\describe{
		\item{X10}{The name of the bin: Period and number.}
		\item{Ocean}{The primary state of the oceans from the point of carbonate precipitation. \code{ar} indicates aragonitic, \code{cc} indicates calcitic conditions.}
		\item{Climate}{Primary climatic characteristic: \code{w} denotes warm, \code{c} denotes cold.}
		\item{\code{bottom}}{Numeric ages of the bottom boundaries (earliest ages) of the bins.}
 	\item{\code{mid}}{Numeric ages midpoints of the bins, the averages of \code{bottom} and \code{top}.}
 	\item{\code{top}}{Numeric ages of the tops (latest ages) of the bins.}
	\item{\code{dur}}{Numeric ages of the durations of the bins.}
 	\item{\code{ten}}{Integer number identifiers of the bins. §correct to num!}
}
}
\source{
Executive committee meeting (2015) of old Paleobiology Database. Additional variables were added by Wolfgang Kiessling.
}
\usage{
data(tens)
}
\description{
Roughly 10 million year timescale used in some analyses.
}
\details{
This is an example time scale object that can be used in the Phanerozoic scale analyses. This time scale comprises 49 bins, roughly 10 million years of durations that result from the combination of certain standard stages.
}
\keyword{datasets}
