% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10.R
\name{top_coal}
\alias{top_coal}
\title{Show the top 10 Coal producers of the given year}
\usage{
top_coal(x)
}
\arguments{
\item{x}{A numerical value between 1980 and 2021.}
}
\value{
A tibble which shows the top 10 coal producers for that year if x is
between 1980 and 2021 else it will return an error message.
}
\description{
This function takes an input between 1980 and 2021 and returns the
top 10 coal producers of the given year
}
\examples{
top_coal(2005)
top_coal(1990)
top_coal(1980)


}
