% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10.R
\name{top_oil}
\alias{top_oil}
\title{Show the top 10 oil producers of the given year}
\usage{
top_oil(x)
}
\arguments{
\item{x}{A numerical value between 1980 and 2021.}
}
\value{
A tibble which shows the top 10 oil producers for that year if x is
between 1980 and 2021 else it will return an error message.
}
\description{
This function takes an input between 1980 and 2021 and returns the
top 10 oil producers of the given year
}
\examples{
top_oil(2005)
top_oil(1990)
top_oil(1980)


}
