% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimws_df.R
\name{trimws_df}
\alias{trimws_df}
\title{Removes leading and trailing whitespace from a dataframe}
\usage{
trimws_df(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{passes arguments from trimws}
}
\value{
dataframe
}
\description{
trimws only removes leading and trailing whitespace from a vector. trimws_df is an extension that will remove  whitespace in an entire dataframe.
}
