% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimwsdf.R
\name{trimwsdf}
\alias{trimwsdf}
\title{Removes leading and trailing whitespace from a dataframe}
\usage{
trimwsdf(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
trimws only removes leading and trailing whitespace from a vector. trimwsdf is an extension that will remove leading and trailing whitespace in all columns in a dataframe. One day I'll build in the same arguments from trimws into trimwsdf. Function created by MarkusN here: https://stackoverflow.com/a/37815274.
}
