% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_screening.R
\name{variable_pre_screening}
\alias{variable_pre_screening}
\title{Variable Screening for Top Predictors}
\usage{
variable_pre_screening(X, y, K = NULL)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. Can be continuous and normally distributed or binary responses.}

\item{K}{Number of top predictors to select. If not specified, default is ncol(X)/2.}
}
\value{
A matrix containing the selected predictor variables based on statistical significance.
}
\description{
This function performs Univariate filtering (Univariate filtering evaluates each feature's individual association with
the target variable through statistical tests, selecting the most significant features based on measures such as correlation,
ANOVA, or chi-squared tests.) on each column of the predictor matrix X based on the provided output variable y.
Various statistical tests are applied depending on the data types of X and y to identify significant predictors.
If y is numeric and X is numeric, Pearson's correlation test or Spearman's rank correlation test is applied.
If y is numeric and X is binary, ANOVA or Kendall's rank correlation test is performed.
If y is binary and X is numeric, ANOVA or Fisher's Exact Test is used.
If both y and X are binary, Chi-squared test is applied.
}
\examples{
# Example usage:
X <- matrix(rnorm(100 * 25), ncol = 25)
y <- rbinom(100, 1, 0.5)
selected_vars <- variable_pre_screening(X, y, 10)

}
\seealso{
\code{pScore} for calculating the statistical significance of predictors.
}
