% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable pre screening.R
\name{variable_screening}
\alias{variable_screening}
\title{Variable Screening for top K predictors}
\usage{
variable_screening(X, y, K = NULL)
}
\arguments{
\item{X}{matrix of the predictor variables}

\item{y}{matrix of the output variables}

\item{k}{Value of number of predictors to be selected
If not specified this value will be nCol(X)/2}
}
\value{
X matrix of the selected variables
}
\description{
This function performs univariate filtering on each column of variable
matrix X and performs various tests according to the combinations of x
and y variables.
In this scenario, we are only considering binary variables under the categorical
category.
If y is numeric, and x is numeric - Pearson's Test(Linear) or Spearman's Test(Non-Linear)
If y is numeric, and x is binary - ANOVA or Kendall's
If y is binary and x is numeric - ANOVA or Kendall's
If y is binary and x is binary - Chi Squared
}
