% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{verify_config}
\alias{verify_config}
\title{This function verifies that all required config fields are provided.}
\usage{
verify_config(config)
}
\arguments{
\item{config}{a config object}
}
\value{
Returns TRUE for a valid config, FALSE otherwise, in which case a list of
missing parameters will be printed out as well
}
\description{
This function verifies that all required config fields are provided.
}
\examples{
##TODO##

# add example
# verify_config(config = NA)
}
