% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{violin_plot}
\alias{violin_plot}
\title{Create a violin plot with consistent formatting}
\usage{
violin_plot(data, x, y, title, xlab = NULL, ylab = NULL)
}
\arguments{
\item{data}{The dataframe containing the data to plot.}

\item{x}{The variable to use for the x-axis.}

\item{y}{The variable to use for the y-axis.}

\item{title}{The title of the plot.}

\item{xlab}{(Optional) The label for the x-axis. Defaults to the variable name of \code{x}.}

\item{ylab}{(Optional) The label for the y-axis. Defaults to the variable name of \code{y}.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function creates a violin plot with consistent formatting, including filling the plot by the x argument (based on the levels of the factor) and defaulting x and y labels to respective variable names.
}
\examples{
violin_plot(mtcars, gear, mpg, "MPG Distribution by Gear")
}
