% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{visualizes the matching/filling procedures}
\usage{
vis(object, column = NULL, details = FALSE)
}
\arguments{
\item{object}{Object returned by the \code{fill} function.}

\item{column}{The empty column for which we want to visualize the matching/filling procedures. By default the
first empty column is taken.}

\item{details}{Displays \code{id}, \code{column}, \code{class} and \code{feature} elements in the table. They are
not displayed by default.}
}
\value{
list containing the pairs of value code and label
}
\description{
The function aims at simplyfying.
}
\examples{
\dontrun{
 #rdf is an object returned by fill function
 vis(rdf)
 vis(rdf, details = TRUE)
}

}
