% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_db_map.R
\name{visualize.DB.map}
\alias{visualize.DB.map}
\title{Visualize Database Map}
\usage{
visualize.DB.map(db_map, layout = "kk")
}
\arguments{
\item{db_map}{A data.table object created by `create.DB.map`.}

\item{layout}{The layout algorithm to use for the graph. Default is "kk" (Kamada-Kawai).
Other options include "fr" (Fruchterman-Reingold), "circle", "grid", etc.}
}
\value{
A graph visualization of the database map.
}
\description{
Generate a visualization of the database map created by `create.DB.map`.
The visualization shows the relationships between tables, including linking variables,
join types, and potential aggregations.
}
\examples{
library(data.table)
table1 <- data.table(id = c(1, 2), value1 = c("A", "B"))
table2 <- data.table(id = c(2, 3), value2 = c("C", "D"))
tables <- list(table1 = table1, table2 = table2)
by <- list(list(by.x = "id", by.y = "id"))  # Corrected syntax
db_map <- create.DB.map(tables, by, join_types = "inner")
visualize.DB.map(db_map)

}
