% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph dataset}
\format{
A data.frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Starting node of the edge (integer)}
  \item{v2}{Ending node of the edge (integer)}
  \item{w}{Weight of the edge (numeric)}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing a sample graph with edges and weights, used to 
  demonstrate Dijkstra's algorithm.
}
\keyword{datasets}
