% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_trend.R
\name{win_trend}
\alias{win_trend}
\title{Winning Trend of an NFL team}
\source{
See the full team list \href{https://www.nfl.com/teams/}{here}.
}
\usage{
win_trend(team)
}
\arguments{
\item{team}{NFL team name}
}
\value{
A trendline of home and away win over the years for the inputted NFL Team.
}
\description{
This function takes the name of an NFL team and plots the winning trend of that team.
The trend will span across 2000 until 2019, with separate lines indicating wins at home and away.
}
\examples{
# Example Input
"Arizona Cardinals"

# Plot the winning trend
win_trend("Arizona Cardinals")


}
