% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{write_config_skeleton}
\alias{write_config_skeleton}
\title{Writes out a config skeleton}
\usage{
write_config_skeleton(file_path = "./config_skeleton.R",
  overwrite = FALSE)
}
\arguments{
\item{file_path}{file path to write the file into}

\item{overwrite}{overwrite existing file defaults to FALSE}
}
\value{
returns a boolean indicating success or failure
}
\description{
Writes out a config skeleton
}
\details{
This function writes out a config skeleton, that is, an empty config file to be edited by the user.
}
\examples{
# set config_empty.R file path
config_file_path <- file.path(tempdir(), "config_empty.R")
#writes out a config skeleton
write_config_skeleton(config_file_path)
}
