% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_out.R
\name{write_csv_out}
\alias{write_csv_out}
\title{write_csv_out}
\usage{
write_csv_out(
  output,
  file_name = deparse(substitute(output)),
  note = NULL,
  date = TRUE,
  increment = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{output}{the object to be written as a .csv file.}

\item{file_name}{the name of the .csv file. Default is to be the same as the name of the output.}

\item{note}{optional text to append to the file_name. Intended to be a short description of the output supplemental to the \code{file_name}.}

\item{date}{prepend the system date in '\%Y\%m\%d' format. Default \code{TRUE}.}

\item{increment}{if a file with the same name exists, save as 'filename-n.csv' with incremented n. Default \code{TRUE}. Cannot be \code{TRUE} when \code{overwrite} is also \code{TRUE}.}

\item{overwrite}{if a file with the same name exists, overwrite. Default \code{FALSE}. Cannot be \code{TRUE} when \code{increment} is also \code{TRUE}.}

\item{...}{additional \code{\link[utils]{write.csv}} arguments. Default change is \code{row.names = F}.}
}
\description{
Wrapper for \code{\link[utils]{write.csv}} with additional functionality to automatically add the .csv file extension, prepend the date, and generate file overwrite warnings.
}
\examples{
## Not run:

# Will write mtcars.csv
write_csv_out(mtcars, date = FALSE)
# Will increment to mtcars-1.csv
write_csv_out(mtcars, date = FALSE)

# Will write the iris dataset to mtcars.csv
write_csv_out(iris, "mtcars", date = FALSE, overwrite = TRUE)

## End(Not run)
}
