% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlcolconv.R
\name{xlcolconv}
\alias{xlcolconv}
\title{Converts Excel column references}
\usage{
xlcolconv(col)
}
\arguments{
\item{col}{Column number to turn into Excel's 'A1 reference style' for columns. Alternatively, an 'A1 reference style' string to turn into a column number.}
}
\value{
A string with the corresponding 'A1 reference style' column reference. Alternatively, a number corresponding to the given 'A1 reference style' column reference.
}
\description{
Converts Excel 'A1 reference style' column references to 'numeric' or vice versa.
}
\examples{
xlcolconv(1)
xlcolconv(26)
xlcolconv(27)
xlcolconv("A")
xlcolconv("Z")
xlcolconv("AA")
}
