% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{affinity}
\alias{affinity}
\title{Calculate environmental affinities of taxa}
\usage{
affinity(dat, env, tax = "occurrence.genus_name", bin = "SLC",
  coll = "collection_no", alpha = 1, reldat = NULL)
}
\arguments{
\item{dat}{(data.frame): the occurrence dataset containing the taxa with unknown environmental affinities.}

\item{env}{(char): The name of the column with the occurrences' environmental values.}

\item{tax}{(char): the column name of taxon names.}

\item{bin}{(char): the column name of bin names.}

\item{coll}{(char): the column name of collection numbers.}

\item{alpha}{(num): the alpha value of the binomial tests. By default (alpha=1) binomial testing is off.}

\item{reldat}{(data.frame): database with the same structure as 'dat'. Typically 'dat' is the subset of 'reldat'. If given, the occurrence distribution of 'reldat' is used 
as the null model of sampling.}
}
\description{
This function will return the environmental affinities of taxa, given the sampling conditions implied by the supplied dataset.
}
\details{
Sampling patterns have an overprinting effect on the frequency of taxon occurrences in different environments. The environmental affinity (sensu Kiessling and Aberhan, 2007 - Paleobiology 33, 414-434 and Kiessling and Kocsis, 2015 - Paleobiology 41, 402-414) expresses whether the taxa are more likely to occurr in an environment, given the sampling patterns of the dataset at hand.
}
\examples{
data(scleractinia)
# omit values where no occurrence environment entry is present, or where unknown
fossils<-subset(scleractinia, slc!=95)
fossilEnv<-subset(fossils, bath!="uk")
# calculate affinities
  aff<-affinity(fossilEnv, env="bath", tax="genus", bin="slc", alpha=1)

}
