% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{\code{analyze} submits the dataset to the analyze API endpoint and
returns the proposed structure of the data. At the current stage
dataset must be rectangular.}
\usage{
analyze(
  df = NULL,
  shape = "vertical",
  inits = 1,
  fast = TRUE,
  sep = ",",
  db = FALSE,
  token = getOption("rejustify.token"),
  email = getOption("rejustify.email"),
  url = getOption("rejustify.mainUrl")
)
}
\arguments{
\item{df}{The dataset to be analyzed. Must be matrix-convertible.}

\item{shape}{It informs the API whether the data set should be read in by
columns (vertical) or by rows (horizontal). The default is vertical.}

\item{inits}{It informs the API how many initial rows (or columns in
horizontal data), correspond to the header description. The default
is inits=1.}

\item{fast}{Informs the API on how big a sample of original data should be.
The larger the sample, the more precise but overall slower the algorithm.
Under the \code{fast = TRUE} the API samples 15% of data, under the
\code{fast = FALSE} option it is 50%. Default is TRUE.}

\item{sep}{The header can also be described by single field values,
separated by a given character separator, for instance 'GDP, Austria, 1999'.
The option informs the API which separator should be used to split the
initial header string into corresponding dimensions. The default is ','.}

\item{db}{It is TRUE if the user accepts rejustify to track her/his activity
to enhance the performance of the AI algorithms. The default is FALSE.}

\item{token}{API token. By default read from global variables.}

\item{email}{E-mail address for the account. By default read from global variables.}

\item{url}{API url. By default read from global variables.}
}
\value{
rejustify API response or errors
}
\description{
\code{analyze} submits the dataset to the analyze API endpoint and
returns the proposed structure of the data. At the current stage
dataset must be rectangular.
}
\examples{
analyze(df)

}
