% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{apply_ecology}
\alias{apply_ecology}
\title{apply_ecology is the user provided ecology function}
\usage{
apply_ecology(abundance, traits, local_environment, config)
}
\arguments{
\item{abundance}{a named vector of abundances with one abundance value per species}

\item{traits}{a named matrix containg the species traits, one row per species}

\item{local_environment}{the environemntal values for the given cell}

\item{config}{the current config}
}
\value{
an abundance vector with the new abundace values for every species.
An abundance value of 0 indicates species death, any other values survival.
}
\description{
apply_ecology is the user provided ecology function
}
\details{
The arguments to this function are an aggregation of all species living in a single cell.
It is then responsible to determine which of the species survives in this particular cell and which ones die.
}
