% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{apply_evolution}
\alias{apply_evolution}
\title{apply_evolution let's the user mutate the traits of a given species}
\usage{
apply_evolution(species, cluster_indices, landscape, config)
}
\arguments{
\item{species}{The current species to mutate}

\item{cluster_indices}{an index vector indicating the cluster every occupied cell is part of}

\item{landscape}{the current landscape}

\item{config}{the current config}
}
\value{
the mutated species traits matrix
}
\description{
apply_evolution let's the user mutate the traits of a given species
}
\details{
This function is called for any single species alongside an index for
the geographical clusters within the species.
}
