% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{apply_evolution}
\alias{apply_evolution}
\title{Allows defining the function that changes the values of traits of a given
species at each time-step and in each site. If no operations are provided, traits are not changing}
\usage{
apply_evolution(species, cluster_indices, space, config)
}
\arguments{
\item{species}{the target species object whose traits will be changed}

\item{cluster_indices}{an index vector indicating the cluster every occupied site is part of}

\item{space}{the current space which can co-determine the rate of trait changes}

\item{config}{the current config}
}
\value{
the mutated species traits matrix
}
\description{
Allows defining the function that changes the values of traits of a given
species at each time-step and in each site. If no operations are provided, traits are not changing
}
\details{
This function is called for any single species alongside an index for
the geographical clusters within the species
}
