% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_var_over_repetition.R
\name{average_RT_participants}
\alias{average_RT_participants}
\title{Plot average RT of all participants over repetitions}
\usage{
average_RT_participants(
  data,
  xlim = NULL,
  ylim = NULL,
  filepath = "../Figures"
)
}
\arguments{
\item{data}{A data frame. NA values will be removed before plotting.}

\item{xlim}{A vector of 2 (for example: c(0, 10)), indicating the range of
the x-axis.If NULL the default value is used.}

\item{ylim}{A vector of 2 (for example: c(0, 1000)), indicating the range of
the y-axis.If NULL the default value is used: c(min(RT), mean(RT) +
(SD(RT))).}

\item{filepath}{A relative or explicit path where plots will be saved}
}
\value{
Plot of RT over repetition, with every line a participant and a pdf file in filepath
}
\description{
Plots the average RT of every session over the repetitions of the facts.
}
\details{
Assumes that the data set has a repetition column with fact repetition in
integers
}
\seealso{
Other average functions: 
\code{\link{average_ROF_facts}()},
\code{\link{average_ROF_participants}()},
\code{\link{average_accuracy_participants}()}
}
\concept{average functions}
