% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{CAGR}
\alias{CAGR}
\title{Calculate compounded growth rate}
\usage{
CAGR(t, `x_i+t`, x_i)
}
\arguments{
\item{t}{Number of compounding periods.}

\item{x_i+t}{Ending value.}

\item{x_i}{Starting value.}
}
\value{
numeric
}
\description{
Calculates the compounded growth rate for between a starting and ending value given a number of (equal) time periods. Usually used to compute the Compound Annual Growth Rate (CAGR) but can be used with different periods.
}
\details{
Computes formula: (x_(i+t)/x_i)^(1/t)-1
}
