% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color_richness}
\alias{color_richness}
\title{Define gen3sis richness color scale}
\usage{
color_richness(n)
}
\arguments{
\item{n}{corresponds to the \link{colorRampPalette} parameter}
}
\value{
returns a \link{colorRampPalette} function with the gen3sis richness colors
}
\description{
Define gen3sis richness color scale
}
