% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crDD.r
\name{crDD}
\alias{crDD}
\title{Diversity dynamics with classical rarefaction (obsolete - will be omitted in version 0.5)}
\usage{
crDD(dat, quota, tax = "genus", bin = "stg", coll = "collection_no",
  iter = 100, noNAStart, average = "geom", intactBins = NULL)
}
\arguments{
\item{dat}{(data.frame): the data frame containing PBDB occurrences.}

\item{quota}{(num): the number of occurrences to rarefy (quota). Slices with lower number of occurrences will be dropped.}

\item{tax}{(char): variable  name of the occurring taxa (variable type: factor) - such as "occurrence.genus_name"}

\item{bin}{(char): variable name of the time slice numbers of the particular occurrences (variable type: int)- such as "stg" or whatever. Bin numbers should be in ascending order,can contain NA's, it can start from a number other than 1 and must not start with 0.}

\item{coll}{(char): variable name of the collection ids.}

\item{iter}{(num): the number of resampling iterations.}

\item{noNAStart}{(bool): useful when the dataset does not start from bin No. 1. Then noNAStart=TRUE will cut the first part of the resulting table, 
so the first row will contain the estimates for the lowest bin number.}

\item{average}{(char): the way the counts are averaged after the trials are finished. Can be 'geom' or 'arit' meaning geometric or arithmetic averaging, respecitvely.}

\item{intactBins}{(num): Vector of those bin identifiers that must be included in the metric calculations, but should not be subsampled. Can be useful for including recent species list, without collection identifiers.}
}
\description{
This script calculates time series of diversity dynamics with iterated subsampling in every time slice.
}
\details{
The output variables are the same as in the divDyn function. Likely updates: c++ core, arguments: sd - list output.
}
\examples{
# import data
  data(corals)
  data(stages)

# calculate metrics of diversity dynamics
   dd <- divDyn(corals, tax="genus", bin="stg")
  rarefDD <- crDD(corals, quota=100, iter=10, intactBins=95)

# plotting
  tsplot(stages, shading="series", boxes="per", xlim=52:95, 
  ylab="range-through diversity (genera)", ylim=c(0,230))
  lines(stages$mid, dd$divRT, lwd=2)
  lines(stages$mid,rarefDD$divRT, col="blue", lwd=2)
  legend("topleft", legend=c("raw","rarefaction"),
    col=c("black", "blue"), lwd=c(2,2), bg="white")

}
