% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{create_config}
\alias{create_config}
\title{The function create_config creates either an empty configuration or a pre-filled configuration from a config file.}
\usage{
create_config(config_file = NA)
}
\arguments{
\item{config_file}{if NA it creates an empty config or the path to a valid configuration file}
}
\description{
The function create_config creates either an empty configuration or a pre-filled configuration from a config file.
}
\examples{
##TODO##

# add example
# create_config(config_file = NA)
}
