% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityDynamics.r
\name{divDyn}
\alias{divDyn}
\title{Diversity dynamics function}
\usage{
divDyn(dat, tax = "occurrence.genus_name", bin = "bin", noNAStart = F,
  inf = F)
}
\arguments{
\item{dat}{(data.frame): the data frame containing PBDB occurrences.}

\item{tax}{(char): variable  name of the occurring taxa (variable type: factor) - such as "occurrence.genus_name"}

\item{bin}{(char): variable name of the time slice numbers of the particular occurrences (variable type: int)- such as "slc" or whatever. Bin numbers should be in ascending order,can contain NA's, it can start from a number other than 1 and must not start with 0.}

\item{noNAStart}{(bool): useful when the dataset does not start from bin No. 1. Then noNAStart=TRUE will cut the first part of the resulting table, 
so the first row will contain the estimates for the lowest bin number.}

\item{Inf}{(logical): should infinites be converted to NAs?}
}
\description{
This script contains mostly the work of Wolfgang Kiessling. It calculates 
various aspects of a time series resolved from PBDB occurrences.
}
\details{
The following variables are produced:

bin: the time slice number

tThrough: the number of through-ranging taxa

tOri: the number of originating taxa

tExt: the number of taxa getting extinct

t1: the number of stratigraphic singleton (single-interval) taxa

t2d: the number of taxa that are present in the i-1th and the ith interval (lower two timers)

t2u: the number of taxa that are present in the ith and the i+1th interval (upper two timers)

tGFu: upper gap-fillers ($later)

tGFd: lower gap-fillers ($later)

t3: the number of three timer taxa, present in time slice i-1, i, and i+1

tPart: the part timer taxa, present in time slice i-1,and i+1, but not in i

extPC: the per capita extinction rates of foote (not normalized with interval length!)

oriPC: the per capita origination rates of foote (not normalized with interval length!)

ext3t: the three-timer extinction rates

ori3t: the three-timer origination rates

extC3t: the corrected three-timer extinction rates

oriC3t: the corrected three-timer origination rates

divSIB: sampled-in-bin diversity (richness)

divCSIB: corrected sampled-in-bin diversity (richness)

divBC: boundary-crosser diversity (richness)

divRT: range-through diversity (richness)

sampRange: sampling probability (Foote)

samp3t: three-timer sampling completeness (used as a correcting value in nC3tExt, nC3tOri and nCorrSIB)

extGF: gap-filler extinction rates

oriGF: gap-filler origination rates
}
\examples{
# import data
  data(scleractinia)
  data(stages)

# calculate metrics of diversity dynamics
   dd <- divDyn(scleractinia, tax="genus", bin="slc")

# plotting
  library(geoscale)
  geoscalePlot(stages$mid,dd$divRT, age.lim=c(260,0), 
    data.lim=c(0,300), type="l", units="Period", vers="ICS2014", 
    ts.width=0.1, tick.scale=50, boxes="Period",ts.col=F,
    label="range-through diversity (genera)")

}

