% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLad.r
\name{fadLad}
\alias{fadLad}
\title{FAD - LAD matrix from occurrence data}
\usage{
fadLad(dat, tax, bin, maxdate = F, mindate = F)
}
\arguments{
\item{dat}{(data.frame): Occurrence data.}

\item{tax}{(char): the column name of taxon names.}

\item{bin}{(char): the column name of bin names.}

\item{maxdate}{(char): optional, the column names of maximum radiometric dates of occurrences.}

\item{mindate}{(char): optional, the column names of minimum radiometric dates of occurrences.}
}
\description{
Function to generate range data from an occurrence dataset.
}
\details{
The following variables are produced:

taxon: the name of the taxon.

FAD: first appearance date in the given time slices.

LAD: last appearance date in the given time slices.

range: the range of the taxo in the given time slices

FAD_DATE: optionally, first appearance dates in numeric ages.

LAD_DATE: optionally, last appearance dates in numeric ages.

DUR: optionally, the duration of taxa in numeric ages. It is given for single-interval taxa as well assuming that the taxa's range span over the entire time slice.

nThrough: the number of through-ranging taxa
}
\examples{
data(scleractinia)
fl <- fadLad(scleractinia, tax="genus", bin="slc", maxdate="ma_max", mindate="ma_min")
}
