% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_species.R
\name{get_divergence_matrix}
\alias{get_divergence_matrix}
\title{get_divergence_matrix returns the full genetic distance matrix for a given species (cell x cell)}
\usage{
get_divergence_matrix(species)
}
\arguments{
\item{species}{the species to produce the full genetic distance matrix for}
}
\value{
the full/decompressed divergence matrix
}
\description{
get_divergence_matrix returns the full genetic distance matrix for a given species (cell x cell)
}
