% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get_geo_richness}
\alias{get_geo_richness}
\title{calculate the richness of a list of species over a given landscape}
\usage{
get_geo_richness(species_list, landscape)
}
\arguments{
\item{species_list}{a list of species to include in the richness calculations}

\item{landscape}{the landscape to calculate the richnness over}
}
\value{
a vector with the richness for every cell in the input landscape
}
\description{
calculate the richness of a list of species over a given landscape
}
\examples{
##TODO##

# add example
# get_geo_richness(species_list, landscape)

}
