% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{linear_model}
\alias{linear_model}
\title{Linear regression model}
\usage{
linear_model(X, Y)
}
\arguments{
\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{Y}{A vector or matrix (one column only) of outcome values.}
}
\value{
A list of attributes of the linear model.
}
\description{
This function is used to build linear models. It can be used to carry out simple linear regression (SLR),
multiple linear regression (MLR).
}
\examples{

}
