% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.editPublicAccess}
\alias{path.editPublicAccess}
\title{Update the public access of a path}
\usage{
path.editPublicAccess(
  pathId,
  token,
  accessLevel = NULL,
  hidden = NULL,
  processingUnits = NULL,
  geoFence = NULL
)
}
\arguments{
\item{pathId}{Mandatory (uuid)}

\item{token}{Mandatory (string)}

\item{hidden}{Optional (logical)}

\item{access}{Optional (named list)}
}
\value{
...
}
\description{
Update the public access of a path
}
