% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path.editPublicAccess}
\alias{path.editPublicAccess}
\title{export}
\usage{
path.editPublicAccess(
  pathId,
  token,
  accessLevel = NULL,
  hidden = NULL,
  processingUnits = NULL,
  geoFence = NULL
)
}
\description{
export
}
