% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_raster_timestamp.R
\name{path.raster.timestamp.getRaster}
\alias{path.raster.timestamp.getRaster}
\title{Service we deliver: We return an R raster object}
\usage{
path.raster.timestamp.getRaster(
  pathId,
  timestampId,
  extent,
  style = NULL,
  threads = 1,
  token = NULL,
  showProgress = TRUE,
  epsg = 3857
)
}
\description{
Service we deliver: We return an R raster object
}
