% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plot the environment variable of a given landscape}
\usage{
plot_landscape(landscape)
}
\arguments{
\item{landscape}{the landscape to plot the environment from}
}
\description{
Plot the environment variable of a given landscape
}
