% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_ranges}
\alias{plot_ranges}
\title{Plot species ranges of the given list of species on a landscape}
\usage{
plot_ranges(species_list, landscape, distrurbance = 0, max_sps = 10)
}
\arguments{
\item{species_list}{a list of species to use in the richness calculation}

\item{landscape}{a corresponding landscape object}

\item{distrurbance}{value randomly added to shift each species symbol. Useful to enhance visualization in case of multiple species overlaps}

\item{max_sps}{maximum number of plotted species, not recommended above 20}
}
\description{
Plot species ranges of the given list of species on a landscape
}
\examples{
## plot from saved outputs
# get path containing example rasters
datapath <- system.file(file.path("extdata", "WorldCenter"), package="gen3sis")
# get species at t0
species_t_300 <- readRDS(file.path(datapath, 
                        "output/config_worldcenter/species/species_t_300.rds"))
# get landscape at t0
landscape_t_300 <- readRDS(file.path(datapath, 
                          "output/config_worldcenter/landscapes/landscape_t_300.rds"))
# plot range
plot_ranges(species_t_300, landscape_t_300)


# get species at t0
species_t_150 <- readRDS(file.path(datapath, 
                                   "output/config_worldcenter/species/species_t_150.rds"))
# get landscape at t0
landscape_t_150 <- readRDS(file.path(datapath, 
                                     "output/config_worldcenter/landscapes/landscape_t_150.rds"))

# plot ranges at intermediate time-step
plot_ranges(species_t_150, landscape_t_150, distrurbance = 1.5, max_sps = 9)






## plot from within observer
# call plot_richness from inside the end_of_timestep_observer function 
# at the config file:
\dontrun{
plot_ranges(data$all_species, data$landscape)
} 
}
