% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RT_over_time.R
\name{plot_RT_over_time}
\alias{plot_RT_over_time}
\title{Plot RT over time}
\usage{
plot_RT_over_time(data)
}
\arguments{
\item{data}{data frame}
}
\value{
data frame
}
\description{
\code{plot_RT_over_time} plots the RT of one person of one session over time
}
