% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_space_overview}
\alias{plot_space_overview}
\title{Plot the outline of a given space over time}
\usage{
plot_space_overview(space, slices = 2, start_end_times = NULL)
}
\arguments{
\item{space}{the input space to be plotted}

\item{slices}{the amount of slices though time between start and end (default value is 2).}

\item{start_end_times}{the stating and ending times of the simulation (default is NULL, takes the oldest and most recent available)}
}
\value{
no return value, called for plot
}
\description{
Plot the outline of a given space over time
}
