% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_species_presence}
\alias{plot_species_presence}
\title{plot a species' presence on a given landscape}
\usage{
plot_species_presence(species, landscape)
}
\arguments{
\item{species}{a single species}

\item{landscape}{a landscape}
}
\description{
plot a species' presence on a given landscape
}
