% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot simulation default summary}
\usage{
plot_summary(output, sumary_legend = NULL)
}
\arguments{
\item{output}{tsgen3sis output object resulting from a gen3sis simulation}

\item{sumary_legend}{either a satring with _\_n for new lines or NULL. If NULL, provides default summary and simulation information.}
}
\description{
Plot simulation default summary
}
