% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot simulation default summary object}
\usage{
plot_summary(output, summary_title = NULL, summary_legend = NULL)
}
\arguments{
\item{output}{a sgen3sis output object resulting from a gen3sis simulation (i.e. run_simulation)}

\item{summary_title}{summary plot title as character. If NULL, title is computed from input name.}

\item{summary_legend}{either a string using \\n for new lines or NULL. If NULL, provides default summary and simulation information.}
}
\value{
no return value, called for plot
}
\description{
Plot simulation default summary object
}
\examples{
\dontshow{
  #TODO change this to a working example
}

\dontrun{
  # load existing summary example
  datapath <- system.file(file.path("extdata", "WorldCenter"), package = "gen3sis")
  output <- readRDS(file.path(datapath, "output/config_worldcenter/sgen3sis.rds"))
  # plot output summary
  plot_summary(output)
  
  plot_summary(output, summary_title="Example")
  
  ## run simulation and plot summary
  # get path or correct input objects
  datapath <- system.file(file.path("extdata", "CaseStudy1"), package="gen3sis")
  # run simulation and store summary object to output
  output <- run_simulation(config = file.path(datapath,"config/config_fast.R"), 
                           space = file.path(datapath,"space"),
                           output_directory = tempdir())
  # plot output summary
  plot_summary(output)
}
}
\seealso{
\code{\link{run_simulation}}
}
