% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTS}
\alias{plotTS}
\title{Time series plotting using a custom time scale}
\usage{
plotTS(...)
}
\arguments{
\item{...}{Arguments of the \code{\link{tsplot}} function.}
}
\description{
Obsolete name for the \code{\link{tsplot}} function.
}
\details{
Obsolete name for the \code{\link{tsplot}} function.
}
