% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTS.R
\name{plotTS}
\alias{plotTS}
\title{Time series plotting using a custom timescale object}
\usage{
plotTS(tsdat, boxes, ylim = c(0, 1), xlim = NULL, prop = 0.05,
  gap = 0.005, bottom = "bottom", mid = "mid", top = "top",
  xlab = "age (Ma)", ylab = "", shading = NULL, shading.col = c("white",
  "gray80"), plot.args = NULL, boxes.args = NULL, labels.args = NULL)
}
\arguments{
\item{tsdat}{(data frame): time scale data frame}

\item{boxes}{(character): column name indicating the names that should be plotted as a timescale}

\item{ylim}{(numeric): the vertical extent of the plot, analogous to the same argument of plot(). By default it is set to the [0,1] interval.}

\item{xlim}{(numeric): the horizontal extent of the plot, analogous to the same argument of plot(). By default it is set to  plot the entire table. If a numeric vector of length=2 is supplied, it will be interpreted as the standard xlim argument and the plot will be displayed based on numerically constrained ages. If it an integer vector with length>2, the interval corresponding to the row indices of the table will be plotted.}

\item{prop}{(numeric value): proportion of the vertical extent of the plot for the timescale}

\item{gap}{(numeric value): proportion of the vertical extent of the plot that should be a gap betwen the timescale and the plot.}

\item{bottom}{(character): column name of the table for the variable that contains the older ages of intervals.}

\item{mid}{(character): column name of the table for the variable that contains the mean ages of intervals.}

\item{top}{(character): column name of the table for the variable that contains the earliest ages of intervals.}

\item{shading}{(character): column name used for the shading.}

\item{shading.col}{(character): name of colors that will be used for the shading, if shading is set.}

\item{plot.args}{(list): list of arguments that will be passed to the main plot() function.  Can be useful for the suppression of axes, font change etc.}

\item{boxes.args}{(list): list of arguments that will be passed to the rect() function that draws the rectangles of time intervals.}

\item{labels.args}{(list): list of arguments that will be passed to the text() function that draws the labels.}
}
\description{
This function allows the user to quickly plot a time scale data table
}
\details{
As most analysis use an individually compiled time scale object, in order to ensure compatibility between the analyzed and plotted values, the time scale table used for the analysis should be plotted rather than a standardized table.
}
\examples{
data(stages)
  plotTS(stages, boxes="period", shading="series")

# only the Mesozoic, custom axes
  plotTS(stages, boxes="period", shading="stage", xlim=52:81, 
    plot.args=list(axes=F, main="Mesozoic"))
  axis(1, at=seq(250, 75, -25), labels=seq(250, 75, -25))
  axis(2)

# only the Triassic, use the supplied abbreviations
  plotTS(stages, boxes="short", shading="stage", xlim=c(250,199), 
    ylab="variable", labels.args=list(cex=1.5, col="blue"), 
    boxes.args=list(col="gray95"))
}
