% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puissance.R
\name{power}
\alias{power}
\title{Pump power}
\usage{
power(rho, Q, HMT, nu)
}
\arguments{
\item{rho}{density of the fluid}

\item{Q}{volumetric flow rate}

\item{HMT}{Overall pressure losses}

\item{nu}{dynamic viscosity}
}
\value{
the pomp power
}
\description{
pump power for flow in a pipe with friction losses
}
\examples{
power(1000,5,20,12)
81750

}
