% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_handling.R
\name{prepare_directories}
\alias{prepare_directories}
\title{prepare_directories checks if the directories exist, and otherwise creates them.
This function will be called by the simulation, but is made available if the directories should be created
beforehand, for example to redirect the stdout to a file in the output directory.}
\usage{
prepare_directories(
  config_file = NA,
  input_directory = NA,
  output_directory = NA
)
}
\arguments{
\item{config_file}{file path to config file, if NA default config will be used}

\item{input_directory}{path to input directoy, if NA will be derived from config file path}

\item{output_directory}{path to output directory, if NA will be derived from config file path}
}
\value{
returns a named list with the paths for the input and uoutput directories
}
\description{
prepare_directories checks if the directories exist, and otherwise creates them.
This function will be called by the simulation, but is made available if the directories should be created
beforehand, for example to redirect the stdout to a file in the output directory.
}
