% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reynolds number.R
\name{Re}
\alias{Re}
\title{Reynolds number calculation}
\usage{
Re(rho, V, L, nu)
}
\arguments{
\item{rho}{density of the fluid}

\item{V}{flow velocity}

\item{L}{caracteristique length}

\item{nu}{dynamic viscosity}
}
\value{
The value of the Reynolds number
}
\description{
Calculate the Reynolds number for a fluid flowing in a pipe
and determine its flow regime.
}
\examples{

Re(1000,2,0.1,0.001)
Turbulent flow
20000
}
