% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read dataset}
\usage{
read_dataset(file)
}
\arguments{
\item{file}{A csv file.}
}
\value{
data frame
}
\description{
\code{read_dataset} reads the raw data set into a data frame with correct
settings
}
\examples{

}
