% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{\code{register} sets the token and email as global variables
to be accessed by rejustify package. The email must correspond to
the token that was assigned to it.}
\usage{
register(token = NULL, email = NULL)
}
\arguments{
\item{token}{API token.}

\item{email}{E-mail address for the account.}
}
\value{
errors only
}
\description{
\code{register} sets the token and email as global variables
to be accessed by rejustify package. The email must correspond to
the token that was assigned to it.
}
\examples{
register(token = "your_token", email = "your_email")

}
