% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch NFL Shiny App}
\usage{
run_app()
}
\description{
This function will launch a shiny app that shows some preliminary facts about NFL. This is an interactive web application, you can play around with the plot. For example, you can change the filter provided in the application, and the plot will change accordingly.
}
