% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_main.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation}
\usage{
run_simulation(
  config_file = NA,
  config = NA,
  input_directory = NA,
  output_directory = NA,
  timestep_restart = NA,
  save_state = NA,
  save_intermediate_results = NA,
  enable_gc = F,
  verbose = 1
)
}
\arguments{
\item{config_file}{configuration file for the simulation}

\item{config}{configuration object derived from a config file}

\item{input_directory}{directory where the all_geo_hab and distance_matrices reside}

\item{output_directory}{directory for the simulation output}

\item{timestep_restart}{restart an exisitng simulation from this timestep or from the latest timestep}

\item{save_state}{save the internal state of the simulation for restarts}

\item{save_intermediate_results}{save the eco, gen, and geo info}

\item{enable_gc}{enable gc in case of memory shortages}

\item{verbose}{integer value (0, 1 ,2 or 3). If 0 no printed statement, 1 timestep progress, 2 enable additional progress outputs regarding current timestep, 3 aditional information from within modules (default is 1)}
}
\value{
no return value
}
\description{
Run a simulation
}
\examples{
1 + 1 
# have to add help here
}
