% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3sis_main.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation}
\usage{
run_simulation(config = NA, landscape = NA, output_directory = NA,
  timestep_restart = NA, save_state = NA, call_observer = "all",
  enable_gc = F, verbose = 1)
}
\arguments{
\item{config}{configuration file for the simulation or configuration object derived from a config file}

\item{landscape}{directory where the all_geo_hab and distance_matrices reside}

\item{output_directory}{directory for the simulation output}

\item{timestep_restart}{= set the start time timestep. If NA start at the beginning, If "ti" start from the last available timestep, if a number "x" start from timestep x}

\item{save_state}{= save the internal state of the simulation for restarts. If "all" save all timestep, if a vector, saves the desired timesteps if "last", saves only last timestep}

\item{call_observer}{call observer functions if any, NA calls at the start and end times, "all" call all timesteps, You can also provide the timesteps for the observer funciton to be called}

\item{enable_gc}{enable gc in case of memory shortages}

\item{verbose}{integer value (0, 1 ,2 or 3). If 0 no printed statement, 1 timestep progress, 2 enable additional progress outputs regarding current timestep, 3 aditional information from within modules (default is 1)}
}
\value{
a summary object containing a minimal summary on simulation and dynamics progress (alive, speciations, extinctions) and some usefull simulation data
}
\description{
Run a simulation
}
\examples{
## running a simple simulation
library(gen3sis)

#get correct path or correct input objects
datapath <- system.file(file.path("extdata", "CaseStudy1"), package="gen3sis")

#run simulation
r <- run_simulation(config = file.path(datapath,"config/config_rte.R"), 
                landscape = file.path(datapath,"landscape"))
}
