% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_abundance}
\alias{save_abundance}
\title{This function can be called within the observer function to save the species abundances.}
\usage{
save_abundance()
}
\description{
This function can be called within the observer function to save the species abundances.
}
\examples{
\dontrun{
  ## save abundances from within observer
  # this functions should be called inside the end_of_timestep_observer function at the config file:
  save_abundance()
}
}
\seealso{
\code{\link{save_species}}
}
