% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_landscape}
\alias{save_landscape}
\title{saves the current landscape, can be called independantly by the user and is called by 
other observer functions relying on the landscape to be present}
\usage{
save_landscape()
}
\description{
saves the current landscape, can be called independantly by the user and is called by 
other observer functions relying on the landscape to be present
}
