% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_richness}
\alias{save_richness}
\title{This function can be called within the observer function to save the current richness pattern}
\usage{
save_richness()
}
\description{
This function can be called within the observer function to save the current richness pattern
}
\examples{

}
