% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{save_traits}
\alias{save_traits}
\title{This function can be called within the observer function to save the species traits.}
\usage{
save_traits()
}
\description{
This function can be called within the observer function to save the species traits.
}
\examples{
##TODO##
}
