% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCurl.R
\name{setCurl}
\alias{setCurl}
\title{\code{setCurl} sets the default configuration for curl calls}
\usage{
setCurl(
  mainUrl = "http://api.rejustify.com",
  proxy = getOption("rejustify.proxy"),
  proxyUrl = getOption("rejustify.proxyUrl"),
  proxyPort = getOption("rejustify.proxyPort")
)
}
\description{
\code{setCurl} sets the default configuration for curl calls
}
\examples{
to be added

}
