% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCurl.R
\name{setCurl}
\alias{setCurl}
\title{sets the default configuration for curl calls}
\usage{
setCurl(
  mainUrl = "http://api.rejustify.com",
  proxy = getOption("rejustify.proxy"),
  proxyUrl = getOption("rejustify.proxyUrl"),
  proxyPort = getOption("rejustify.proxyPort")
)
}
\arguments{
\item{mainUrl}{Main address for rejustify API calls. Default is set to
\code{http://api.rejustify.com}, but depending on the customer needs, the address may change.}

\item{proxyUrl}{Address of the proxy server.}

\item{proxyPort}{Port for communication with the proxy server.}
}
\description{
The command stores the connection details into memory to be easier accessed by
rejustify package.
}
\examples{
#setting up connection through proxy
rejustify::setCurl(proxyUrl = "PROXY_ADDRESS", proxyPort = 8080)

}
