% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perte singuliere.R
\name{sing_losses}
\alias{sing_losses}
\title{Singular pressure losses}
\usage{
sing_losses(K, V)
}
\arguments{
\item{K}{singular loss coefficient}

\item{V}{flow velocity in  the pipe}
}
\value{
Singular pressure losses
}
\description{
Singular head losses in pipe fittings
}
\examples{

sing_losses(0.4,10)
196.2

}
