% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaklog.r
\name{streaklog}
\alias{streaklog}
\alias{whichmaxstreak}
\title{Utility functions for slicing gappy time series}
\usage{
streaklog(x)

whichmaxstreak(x, which = -1)
}
\arguments{
\item{x}{a vector with gaps. Will be converted to a logical vector, with gaps indicated by values of FALSE.}
}
\description{
The function returns where the continuous streaks start and how long they are, which can be used for efficient and flexible subsetting.
}
\details{
The output list of streak() contains the following elements:
starts: the indices where the streaks start.
streaks: the lengths of the individual streaks (number of values).
runs: the number of streaks.

The function whichmaxstreak() will return the indices of those values that are in the longest continuous streak.
}
\examples{
# generate a sequence of values
  b<-40:1
# add some gaps
  b[c(1:4, 15, 19, 23:27)]  <- NA
# the functions
  streaklog(b)
  whichmaxstreak(b)
}

