% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLad.r
\name{survivors}
\alias{survivors}
\title{Estimation of survivorship probabilities}
\usage{
survivors(dat, tax = "genus", bin = "slc", method = "forward",
  noNAStart = FALSE, fl = NULL)
}
\arguments{
\item{dat}{(data.frame): the data frame containing PBDB occurrences.}

\item{tax}{(char): variable  name of the occurring taxa (variable type: factor) - such as "occurrence.genus_name"}

\item{bin}{(char): variable name of the time slice numbers of the particular occurrences (variable type: int)- such as "slc" or whatever. Bin numbers should be in ascending order,can contain NA's, it can start from a number other than 1 and must not start with 0.}

\item{method}{(character value): either "forward" or "backward".}

\item{noNAStart}{(bool): useful when the dataset does not start from bin No. 1. Then noNAStart=TRUE will cut the first part of the resulting table, 
so the first row will contain the estimates for the lowest bin number.}

\item{fl}{(matrix or data.frame). If so desired, the function can be run on an FAD-LAD dataset. (fadLad)}
}
\description{
This function will calculate both forward and backward survivorship probabilities from a given occurrence dataset or FAD-LAD matrix.
}
\examples{
data(corals)
surv<-survivors(corals, tax="genus", bin="slc", method="forward")

# plot
data(stages)
plotTS(stages, shading="series", boxes="per", xlim=c(260,0), 
  ylab="proportion of survivors present", ylim=c(0.01,1),plot.args=list(log="y"))
  
for(i in 1:ncol(surv)) lines(stages$mid, surv[,i])

}
