% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Add title
Visualization Function}
\usage{
visualize(x)
}
\arguments{
\item{x}{A numeric variable.}
}
\value{
Some basic visualization (graphs).
}
\description{
Add description
Creates histogram, boxplot and numeric summary
}
